% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{add_inference}
\alias{add_inference}
\title{Add inference to augsynth object}
\usage{
add_inference(object, inf_type = "conformal", linear_effect = F, ...)
}
\arguments{
\item{object}{augsynth object}

\item{inf_type}{Type of inference algorithm. Options are
\describe{
 \item{"conformal"}{Conformal inference (default)}
 \item{"jackknife+"}{Jackknife+ algorithm over time periods}
 \item{"jackknife"}{Jackknife over units}
 \item{"permutation"}{Placebo permutation, raw ATT}
 \item{"permutation_rstat"}{Placebo permutation, RMSPE adjusted ATT}
}}

\item{linear_effect}{Boolean, whether to invert the conformal
inference hypothesis test to get confidence intervals for a
linear-in-time treatment effect: intercept + slope * time}

\item{...}{Optional arguments for inference, for more details for
each `inf_type` see
      \describe{
       \item{"conformal"}{`conformal_inf`}
       \item{"jackknife+"}{`time_jackknife_plus`}
       \item{"jackknife"}{`jackknife_se_single`}
       \item{"permutation"}{`permutation_inf`}
      }}
}
\description{
Add inference to augsynth object
}
