% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{alphaloss}
\alias{alphaloss}
\title{compute the synthetic controls loss}
\usage{
alphaloss(outcomes, metadata, trt_unit, alpha)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alphas}{regularization parameter}
}
\description{
compute the synthetic controls loss
}
