% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\docType{package}
\name{augsynth-package}
\alias{augsynth-package}
\title{augsynth}
\description{
A package implementing the Augmented Synthetic Controls Method
}
\author{
\strong{Maintainer}: Eli Ben-Michael \email{ebenmichael@berkeley.edu}

Authors:
\itemize{
  \item Eli Ben-Michael \email{ebenmichael@berkeley.edu}
}

}
