% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes_plot.R
\name{augsynth_outcomes_plot}
\alias{augsynth_outcomes_plot}
\title{Plot function returning the original level of the outcome variable
for the treated unit and its synthetic counterfactual}
\usage{
augsynth_outcomes_plot(augsynth, measure = c("synth", "average"))
}
\arguments{
\item{augsynth}{Augsynth object to be plotted}

\item{measure}{Whether to plot the synthetic counterfactual or the
raw average of donor units}
}
\description{
Plot function returning the original level of the outcome variable
for the treated unit and its synthetic counterfactual
}
