% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{augsynth_plot_from_results}
\alias{augsynth_plot_from_results}
\title{Plot function for summary function for augsynth}
\usage{
augsynth_plot_from_results(augsynth, inf_type = NULL, ...)
}
\arguments{
\item{...}{Optional arguments}

\item{x}{Summary object}
}
\description{
Plot function for summary function for augsynth
}
