% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{bin_search}
\alias{bin_search}
\title{Wrapper for bin_search_ which generates tolerances to search over}
\usage{
bin_search(start, end, by, feasfunc)
}
\arguments{
\item{start}{Starting value of tolerances}

\item{end}{Ending value of tolerances}

\item{by}{Step size of tolerances}

\item{feasfunc}{Function which returns True if feasible}
}
\description{
Wrapper for bin_search_ which generates tolerances to search over
}
