% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{bin_search_}
\alias{bin_search_}
\title{Perform binary search for the smallest balance which is feasible}
\usage{
bin_search_(eps, feasfunc)
}
\arguments{
\item{eps}{Sorted list of balance tolerances to try}

\item{feasfunc}{Function which returns True if feasible}
}
\value{
smallest tolerance which is feasible
}
\description{
Perform binary search for the smallest balance which is feasible
}
