% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{bin_search_balancer}
\alias{bin_search_balancer}
\title{Find the minimal amount of regularization possible}
\usage{
bin_search_balancer(outcomes, metadata, trt_unit = 1, start, end, by,
  link = c("logit", "linear", "pos-linear"), regularizer = c("l1", "l2",
  "linf"), normalized = TRUE, outcome_col = NULL, cols = list(unit =
  "unit", time = "time", outcome = "outcome", treated = "treated"),
  opts = list())
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{start}{Starting value of tolerances}

\item{end}{Ending value of tolerances}

\item{by}{Step size of tolerances}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{normalized}{Whether to normalize the weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error tolerance}}}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Find the minimal amount of regularization possible
}
