% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bootstrap_methods}
\alias{bootstrap_methods}
\alias{rdirichlet_b}
\alias{rmultinom_b}
\alias{rwild_b}
\title{Bootstrap Functions}
\usage{
rdirichlet_b(n)

rmultinom_b(n)

rwild_b(n)
}
\arguments{
\item{n}{Number of units}
}
\description{
There are several helper functions used for bootstrap inference.
Each method returns a list of selection weights of length n,
which weights that sum to n.
}
\section{Functions}{
\itemize{
\item \code{rdirichlet_b()}: Bayesian bootstrap

\item \code{rmultinom_b()}: Non-parametric bootstrap

\item \code{rwild_b()}: Wild bootstrap (Mammen 1993)

}}
