% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bootstrap_ridgeaug}
\alias{bootstrap_ridgeaug}
\title{Bootstrap standard errors for estimated counterfactual with synth}
\usage{
bootstrap_ridgeaug(outcomes, metadata, n_boot, trt_unit = 1, pred_int = F,
  lambda = NULL, scm = T, ridge = T, cols = list(unit = "unit", time =
  "time", outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{n_boot}{Number of bootstrap samples}

\item{trt_unit}{Treated unit}

\item{pred_int}{Whether to add outcome control variance to make pred interval}

\item{lambda}{Ridge hyper-parameter, if NULL then CV}

\item{scm}{Whether to augment SCM with ridge}

\item{ridge}{Include ridge or not}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
att estimates, test statistics, p-values
}
\description{
Bootstrap standard errors for estimated counterfactual with synth
}
