% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{bootstrap_sc}
\alias{bootstrap_sc}
\title{Bootstrap standard errors for estimated counterfactual with synth}
\usage{
bootstrap_sc(outcomes, metadata, n_boot, trt_unit = 1, cols = list(unit =
  "unit", time = "time", outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{n_boot}{Number of bootstrap samples}

\item{trt_unit}{Treated unit}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
att estimates, test statistics, p-values
}
\description{
Bootstrap standard errors for estimated counterfactual with synth
}
