% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{calculate_MDES_table}
\alias{calculate_MDES_table}
\title{Calculate MDES}
\usage{
calculate_MDES_table(lest, treat_year, tx_col, time_col)
}
\arguments{
\item{lest}{Entity by year estimate estimates for all treatment and control entities
(original tx and all the comparisons as pseudo-treated).  Columns
of entity ID (tx_col), treatment flag (trt), time period (time_col),
and estimated tx-synthetic control difference (ATT).}

\item{treat_year}{The time period in which the treatment was introduced.}

\item{tx_col}{The name to the column containing the treatment variable.}

\item{time_col}{The name of the column containing the time variable.}
}
\value{
List of three dataframes, one of MDES, one of RMSPEs, and
  one of info on the r-statistics (the ATTs divided by the RMSPEs).
}
\description{
Use our method for Calculating SEs, p-values, and MDEs by looking
at the distribution of impacts across the donor units.
}
