% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{calculate_RMSPE}
\alias{calculate_RMSPE}
\title{Calculate RMSPE for all units on lagged outcomes}
\usage{
calculate_RMSPE(lest, treat_year, tx_col, time_col)
}
\arguments{
\item{lest}{Entity by year estimate estimates for all treatment and control entities
(original tx and all the comparisons as pseudo-treated).  Columns
of entity ID (tx_col), treatment flag (trt), time period (time_col),
and estimated tx-synthetic control difference (ATT).}

\item{treat_year}{The time period in which the treatment was introduced.}

\item{tx_col}{The name to the column containing the treatment variable.}

\item{time_col}{The name of the column containing the time variable.}
}
\description{
This calculates the RMSPE of the difference between estimated and
observed outcome, averaged across the pre-treatment years, for each
county
}
