% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{check_data_stag}
\alias{check_data_stag}
\title{Check that we can actually run multisynth on the data}
\usage{
check_data_stag(wide, fixedeff, n_leads, n_lags)
}
\arguments{
\item{wide}{Output of format_data_stag}

\item{fixedeff}{Whether to include a unit fixed effect}

\item{n_leads}{How long past treatment effects should be estimated for, default is number of post treatment periods for last treated unit}

\item{n_lags}{Number of pre-treatment periods to balance, default is to balance all periods}
}
\description{
Check that we can actually run multisynth on the data
}
