% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{cherno_test}
\alias{cherno_test}
\title{Weighted Least Squares Standard Errors}
\usage{
cherno_test(outcomes, metadata, ns = 1000, q = c(2, 1), trt_unit = 1,
  permtype = c("iid", "block"), cols = list(unit = "unit", time = "time",
  outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{ns}{Number of samples from the permutation distribution to draw}

\item{trt_unit}{Treated unit}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{order}{of test statistic}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Weighted Least Squares Standard Errors
}
