% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{choose_lambda}
\alias{choose_lambda}
\title{Choose either the lambda that minimizes CV MSE or largest lambda within 1 se of min}
\usage{
choose_lambda(lambdas, lambda_errors, lambda_errors_se, min_1se)
}
\arguments{
\item{lambdas}{list of lambdas}

\item{lambda_errors}{The MSE associated with each lambda term in lambdas.}

\item{lambda_errors_se}{The SE of the MSE associated with each lambda}

\item{min_1se}{If TRUE, chooses the maximum lambda within 1 standard error of the lambda that minimizes the CV error, if FALSE chooses the optimal lambda; default TRUE}
}
\value{
optimal lambda
}
\description{
Choose either the lambda that minimizes CV MSE or largest lambda within 1 se of min
}
