% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{choosealpha}
\alias{choosealpha}
\title{Pick the hyperparameter which gives the best pre-period fit}
\usage{
choosealpha(outcomes, metadata, trt_unit = 1, alphas)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alphas}{regularization parameters to pick from}
}
\value{
Best alpha
}
\description{
Pick the hyperparameter which gives the best pre-period fit
}
