% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{ci95_rstat_plot}
\alias{ci95_rstat_plot}
\title{Generate augsynth plot with 95\% CI based on inference preference}
\usage{
ci95_rstat_plot(augsynth, inf_type)
}
\arguments{
\item{inf_type}{Inference type (takes a value of 'permutation' or 'permutation_rstat')}

\item{results}{Results from calling xxx}
}
\description{
Generate augsynth plot with 95\% CI based on inference preference
}
