% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_basque}
\alias{clean_basque}
\title{Clean up the basque data that comes from the Synth package}
\usage{
clean_basque(df, sim_num = -1)
}
\description{
Clean up the basque data that comes from the Synth package
}
