% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_regression_multi.R
\name{collect_data}
\alias{collect_data}
\title{Organize data for the QP}
\usage{
collect_data(X, trt, n_leads)
}
\arguments{
\item{X}{Matrix of outcomes}

\item{trt}{Vector of treatment status for each unit}

\item{n_leads}{How long past treatment effects should be estimated for}
}
\description{
Organize data for the QP
}
