% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{combine_outcomes}
\alias{combine_outcomes}
\title{Helper function to combine multiple outcomes into a single balance matrix}
\usage{
combine_outcomes(wide_list, combine_method, fixedeff, k = NULL, ...)
}
\arguments{
\item{wide_list}{List of lists of pre/post treatment data for each outcome}

\item{combine_method}{How to combine outcomes}

\item{k}{Number of principal directions to keep, default all}

\item{fixed_eff}{Whether to take out unit fixed effects or not}
}
\value{
\itemize{
         \item{"X"}{Matrix of combined pre-treatment outcomes}
         \item{"trt"}{Vector of treatment assignments}
         \item{"y"}{Matrix of combined post-treatment outcomes}
        }
}
\description{
Helper function to combine multiple outcomes into a single balance matrix
}
