% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compute_att}
\alias{compute_att}
\title{Compute the ATT in the post-period}
\usage{
compute_att(outcomes, metadata, trt_unit = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes}

\item{metadata}{Dataframe with info about the simulation}

\item{trt_unit}{Unit to count as treated, defaults to using
the treated column}
}
\description{
Compute the ATT in the post-period
}
