% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{compute_stat}
\alias{compute_stat}
\title{Compute a test statistic statfunc({Y_t | t in times})}
\usage{
compute_stat(att, pretimes, posttimes, statfunc)
}
\arguments{
\item{att}{ATT estimates in a dataframe}

\item{pretimes}{Vector of times in pre-period}

\item{posttimes}{Vector of times in post-period}

\item{statfunc}{Function to compute test stat, takes in pre and post period}
}
\value{
Value of test statistic
}
\description{
Compute a test statistic statfunc({Y_t | t in times})
}
