% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_outcomes.R
\name{concat_synth_out}
\alias{concat_synth_out}
\title{Fit synthetic controls for multiple outcomes with random weights}
\usage{
concat_synth_out(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 1}
}
\value{
concatenated output of format_synth
}
\description{
Fit synthetic controls for multiple outcomes with random weights
}
