% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_covariate_balance.R
\name{covariate_balance_table}
\alias{covariate_balance_table}
\title{Make covariate balance table.}
\usage{
covariate_balance_table(ascm, pre_period = NULL)
}
\arguments{
\item{ascm}{An augsynth result object from single_augsynth}

\item{pre_period}{List of names of the pre-period timepoints to
calculate balance for.  NULL means none.}
}
\description{
Make a table comparing means of covariates in the treated group,
the raw control group, and the new weighted control group (the
synthetic control)
}
