% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_outcomes.R
\name{create_index}
\alias{create_index}
\title{Create an index of the outcomes with a weighted average}
\usage{
create_index(outcomes, metadata, alpha)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{alpha}{Vector of size (n_outcomes - 1) or n_outcomes
which specifies the weights}
}
\value{
index of outcomes
}
\description{
Create an index of the outcomes with a weighted average
}
