% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{create_lambda_list}
\alias{create_lambda_list}
\title{Create list of lambdas}
\usage{
create_lambda_list(lambda_max, lambda_min_ratio, n_lambda)
}
\arguments{
\item{lambda_max}{Initial (largest) lambda, if NULL sets it to be (1+norm(X_1-X_c))^2}

\item{lambda_min_ratio}{Ratio of the smallest to largest lambda when tuning lambda values}

\item{n_lambda}{Number of lambdas to consider between the smallest and largest lambda value}
}
\value{
List of lambdas
}
\description{
Create list of lambdas
}
