% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{cv_kfold}
\alias{cv_kfold}
\title{Cross validation for l2_entropy regularized synthetic controls
K-fold cross validation, fit on K-1 folds, evaluate balance on Kth fold}
\usage{
cv_kfold(outcomes, metadata, trt_unit, eps, n_folds, outcome_col = NULL,
  lasso = FALSE, cols = list(unit = "unit", time = "time", outcome =
  "outcome", treated = "treated"), max_iters = 1000, tol = 1e-08)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps}{vector of regularization parameters to consider}

\item{n_folds}{Number of CV folds}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{lasso}{Whether to do lasso (every covariate is separate)}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{max_iters}{Maximum number of iterations}

\item{tol}{Convergence tolerance}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Cross validation for l2_entropy regularized synthetic controls
K-fold cross validation, fit on K-1 folds, evaluate balance on Kth fold
}
