% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{cv_kfold_bal}
\alias{cv_kfold_bal}
\title{Cross validation for balancing weights
K-fold cross validation, fit on K-1 folds, evaluate balance on Kth fold}
\usage{
cv_kfold_bal(outcomes, metadata, n_folds, hyperparams, link = c("logit",
  "linear", "pos-linear"), regularizer = c(NULL, "l1", "l2", "ridge", "linf"),
  normalized = TRUE, outcome_col = NULL, cols = list(unit = "unit", time =
  "time", outcome = "outcome", treated = "treated"), opts = list())
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{hyperparams}{Vector of hyper-parameter settings to consider}

\item{regularizer}{Dual of balance criterion}

\item{normalized}{Whether to normalize the weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error tolerance}}}
}
\value{
best hyper parameter
}
\description{
Cross validation for balancing weights
K-fold cross validation, fit on K-1 folds, evaluate balance on Kth fold
}
