% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{cv_lambda}
\alias{cv_lambda}
\title{Choose best lambda with CV}
\usage{
cv_lambda(
  X_c,
  X_1,
  synth_data,
  trt,
  holdout_length,
  scm,
  lambda_max,
  lambda_min_ratio,
  n_lambda,
  min_1se
)
}
\arguments{
\item{X_c}{Matrix of control lagged outcomes}

\item{X_1}{Vector of treated leagged outcomes}

\item{synth_data}{Output of `format_synth`}

\item{trt}{Vector of treatment indicators}

\item{holdout_length}{Length of conseuctive holdout period for when tuning lambdas}

\item{scm}{Include SCM or not}

\item{lambda_max}{Initial (largest) lambda, if NULL sets it to be (1+norm(X_1-X_c))^2}

\item{lambda_min_ratio}{Ratio of the smallest to largest lambda when tuning lambda values}

\item{n_lambda}{Number of lambdas to consider between the smallest and largest lambda value}

\item{min_1se}{If TRUE, chooses the maximum lambda within 1 standard error of the lambda}
}
\value{
\itemize{
         \item{"lambda"}{Value of the ridge hyperparameter}
         \item{"lambdas"}{List of lambda values evaluated to tune ridge regression}
         \item{"lambda_errors"}{"The MSE associated with each lambda term in lambdas."}
         \item{"lambda_errors_se"}{"The SE of the MSE associated with each lambda term}
}
}
\description{
Choose best lambda with CV
}
