% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{cv_wz_bal}
\alias{cv_wz_bal}
\title{Cross validation for l2_entropy regularized synthetic controls
Take bootstrap samples over the controls and evaluate balance
Modified version of Wang & Zubizaretta (2018)}
\usage{
cv_wz_bal(outcomes, metadata, n_boot, hyperparams, link = c("logit", "linear",
  "pos-linear"), regularizer = c(NULL, "l1", "l2", "ridge", "linf"),
  normalized = TRUE, outcome_col = NULL, cols = list(unit = "unit", time =
  "time", outcome = "outcome", treated = "treated"), opts = list())
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{hyperparams}{Vector of hyper-parameter settings to consider}

\item{regularizer}{Dual of balance criterion}

\item{normalized}{Whether to normalize the weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error tolerance}}}
}
\value{
best hyperparameter
}
\description{
Cross validation for l2_entropy regularized synthetic controls
Take bootstrap samples over the controls and evaluate balance
Modified version of Wang & Zubizaretta (2018)
}
