% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{demean_data}
\alias{demean_data}
\title{Remove unit means}
\usage{
demean_data(wide_data, synth_data)
}
\arguments{
\item{wide_data}{X, y, trt}

\item{synth_data}{List with data formatted as Synth::dataprep}
}
\description{
Remove unit means
}
