% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{donor_table}
\alias{donor_table}
\title{Return a summary data frame donor units used in the model, along with their RMSPEs and synthetic weights}
\usage{
donor_table(augsynth)
}
\arguments{
\item{augsynth}{Augsynth object to be plotted}
}
\description{
Return a summary data frame donor units used in the model, along with their RMSPEs and synthetic weights
}
