% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{drop_unit_i}
\alias{drop_unit_i}
\title{Drop unit i from data}
\usage{
drop_unit_i(wide_data, Z, i)
}
\arguments{
\item{wide_data}{(X, y, trt)
param Z Covariates matrix}

\item{i}{Unit to drop}
}
\description{
Drop unit i from data
}
