% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{drop_unit_i_}
\alias{drop_unit_i_}
\title{Helper function to drop unit i from the multisynth object}
\usage{
drop_unit_i_(msyn, i)
}
\description{
Helper function to drop unit i from the multisynth object
}
