% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{drop_unit_i_multi}
\alias{drop_unit_i_multi}
\title{Helper function to drop unit i and refit}
\usage{
drop_unit_i_multi(msyn, i)
}
\arguments{
\item{i}{Unit to drop}
}
\description{
Helper function to drop unit i and refit
}
