% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ents.R
\docType{package}
\name{ents}
\alias{ents}
\alias{ents-package}
\title{ents: A package for maximum entropy synthetic controls with
(possibly) multiple outcomes}
\description{
ents: A package for maximum entropy synthetic controls with
(possibly) multiple outcomes
}
