% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{estimate_robust_SE}
\alias{estimate_robust_SE}
\title{Estimate robust SE}
\usage{
estimate_robust_SE(
  placebo_estimates,
  k = NULL,
  beta = NULL,
  round_beta = FALSE
)
}
\arguments{
\item{placebo_estimates}{The list of impact estimates to use for
estimating the SE.}

\item{k}{Number of obs to drop from each tail (so 2k obs dropped
total)}

\item{beta}{Proportion of obs to drop from each tail. E.g., 95\% to
drop 5\% most extreme.  Only one of beta or k can be non-null.}

\item{round_beta}{TRUE means adjust beta to the nearest integer
number of units to drop (the Howard method).  FALSE means
interpolate quantiles using R's quantile function.}
}
\description{
Given a list of impact estimates (all assumed to be estimating a
target impact of 0), calculate the modified standard deviation by
looking an in inner-quartile range rather than the simple standard
deviation.  This reduces the effect of small numbers of extreme
outliers, and focuses on central variation.  This then gets
converted to a standard error (standard deviation of the results,
in effect.)
}
