% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{firpo_inf}
\alias{firpo_inf}
\title{Get the permutation distribution of the test stat
assuming equal probability of treatment}
\usage{
firpo_inf(metadata, fitfunc, units, trt_unit, pretimes, posttimes, statfuncs)
}
\arguments{
\item{metadata}{with treatment time}

\item{fitfunc}{Partially applied fitting function which takes in
the number of the treated unit}

\item{units}{Numbers of the units to permute treatment around}

\item{trt_unit}{Treated unit}

\item{pretimes}{Vector of times in pre-period}

\item{posttimes}{Vector of times in post-period}

\item{statfuncs}{Function to compute test stats}
}
\value{
att estimates, test statistics, p-values
}
\description{
Get the permutation distribution of the test stat
assuming equal probability of treatment
}
