% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{firpo_inf_synth}
\alias{firpo_inf_synth}
\title{Get the permutation distribution of SC estimate test statistics
assuming equal probability of treatment}
\usage{
firpo_inf_synth(outcomes, metadata, trt_unit, pos = FALSE,
  statfuncs = c(rmse_ratio, mean_abs, abs_tstat), include_treat = FALSE,
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{pos}{Whether to only consider units with positive weights, default False}

\item{statfuncs}{Function to compute test stats}

\item{include_treat}{Whether to include the treated unit in randomization distribution}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
att estimates, test statistics, p-values
}
\description{
Get the permutation distribution of SC estimate test statistics
assuming equal probability of treatment
}
