% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_augsyn}
\alias{fit_augsyn}
\title{Fit outcome model and balance pre-period}
\usage{
fit_augsyn(wide_data, synth_data, progfunc = c("EN", "RF", "GSYN", "MCP",
  "CITS", "CausalImpact", "seq2seq"), weightfunc = c("SC"),
  opts.prog = NULL, opts.weights = NULL)
}
\arguments{
\item{wide_data}{Output of `format_ipw`}

\item{synth_data}{Output of `synth_data`}

\item{progfunc}{What function to use to impute control outcomes
EN=Elastic Net, RF=Random Forest, GSYN=gSynth,
Comp=softImpute, MCP=MCPanel, CITS=CITS
CausalImpact=Bayesian structural time series with CausalImpact
seq2seq=Sequence to sequence learning with feedforward nets}

\item{weightfunc}{What function to use to fit weights
SC=Vanilla Synthetic Controls, ENT=Maximum Entropy}

\item{opts.prog}{Optional options for fitting prognostic score}

\item{opts.weights}{Optional options for fitting synth weights}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
}
}
\description{
Fit outcome model and balance pre-period
}
