% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_augsyn_formatted}
\alias{fit_augsyn_formatted}
\title{Fit E[Y(0)|X] and for each post-period and balance pre-period}
\usage{
fit_augsyn_formatted(wide_data, synth_data, fit_progscore, fit_weights,
  ...)
}
\arguments{
\item{wide_data}{Output of `format_ipw`}

\item{synth_data}{Output of `synth_data`}

\item{fit_progscore}{Function to fit prognostic score}

\item{fit_weights}{Function to fit synth weights}

\item{...}{optional arguments for outcome model}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
}
}
\description{
Fit E[Y(0)|X] and for each post-period and balance pre-period
}
