% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{fit_augsynth_internal}
\alias{fit_augsynth_internal}
\title{Internal function to fit augmented SCM}
\usage{
fit_augsynth_internal(wide, synth_data, Z, progfunc, scm, fixedeff,
  V = NULL, ...)
}
\arguments{
\item{wide}{Data formatted from format_data}

\item{synth_data}{Data formatted from foramt_synth}

\item{Z}{Matrix of auxiliary covariates}

\item{progfunc}{outcome model to use}

\item{scm}{Whether to fit SCM}

<<<<<<< e0c10727454c99539f69fcecab29c50d683a22c4
\item{fixedeff}{Whether to de-mean synth}
=======
\item{V}{V matrix for Synth, default NULL}

\item{...}{Extra args for outcome model}
>>>>>>> More synth for multiple outcomes

\item{...}{Extra args for outcome model}
}
\description{
Internal function to fit augmented SCM
}
