% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{fit_augsynth_multiout_internal}
\alias{fit_augsynth_multiout_internal}
\title{Internal function to fit augmented SCM with multiple outcomes}
\usage{
fit_augsynth_multiout_internal(wide_list, combine_method, Z, progfunc, scm,
  fixedeff, ...)
}
\arguments{
\item{wide_list}{List of matrices for each outcome formatted from format_data}

\item{combine_method}{How to combine outcomes}

\item{Z}{Matrix of auxiliary covariates}

\item{progfunc}{outcome model to use}

\item{scm}{Whether to fit SCM}

\item{fixedeff}{Whether to de-mean synth}

\item{...}{Extra args for outcome model}
}
\description{
Internal function to fit augmented SCM with multiple outcomes
}
