% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancer.R
\name{fit_balancer_formatted}
\alias{fit_balancer_formatted}
\title{Find Balancing weights by solving the dual optimization problem}
\usage{
fit_balancer_formatted(X, trt, link = c("logit", "linear", "pos-linear"),
  regularizer = c(NULL, "l1", "l2", "ridge", "linf"), hyperparam,
  normalized = TRUE, opts = list())
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{hyperparam}{Regularization hyperparameter}

\item{normalized}{Whether to normalize the weights}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error rolerance}}}
}
\value{
\itemize{
         \item{theta }{Estimated dual propensity score parameters}
         \item{weights }{Estimated primal weights}
         \item{imbalance }{Imbalance in covariates}}
}
\description{
Find Balancing weights by solving the dual optimization problem
}
