% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_dr}
\alias{fit_dr}
\title{Fit a regularized outcome model and synthetic controls}
\usage{
fit_dr(outcomes, metadata, trt_unit = 1, alpha_w, alpha_o)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alpha_w}{regularization parameter for weights}

\item{alpha_o}{regularization parameter for outcome model}
}
\value{
synthetic control weights,
        outcome regression parameters
        control outcomes
        boolean for treated
}
\description{
Fit a regularized outcome model and synthetic controls
}
