% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_dr_formatted}
\alias{fit_dr_formatted}
\title{Fit a regularized outcome model and synthetic controls
for a double robust estimator}
\usage{
fit_dr_formatted(data_out, alpha_w, alpha_o, syn = TRUE)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{alpha_w}{regularization parameter for weights}

\item{alpha_o}{regularization parameter for outcome model}

\item{syn}{whether to use synthetic control weights in DR estimate}
}
\value{
synthetic control weights,
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Fit a regularized outcome model and synthetic controls
for a double robust estimator
}
