% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{fit_ebal_formatted}
\alias{fit_ebal_formatted}
\title{Fit entropy balancing weights}
\usage{
fit_ebal_formatted(data_out)
}
\arguments{
\item{data_out}{formatted data from format_ipw}
}
\value{
entropy balancing weights
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Fit entropy balancing weights
}
