% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_entropy}
\alias{fit_entropy}
\title{Fit l2 entropy regularized synthetic controls on outcomes
Wrapper around fit_l2_entropy_formatted}
\usage{
fit_entropy(outcomes, metadata, trt_unit = 1, eps = NULL,
  outcome_col = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps}{Bound on synthetic control differences}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}
}
\value{
Weights for synthetic controls, control outcomes as matrix,
        and whether the unit is actually treated
}
\description{
Fit l2 entropy regularized synthetic controls on outcomes
Wrapper around fit_l2_entropy_formatted
}
