% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_entropy_formatted}
\alias{fit_entropy_formatted}
\title{Fit entropy regularized synthetic controls
by solving the dual problem}
\usage{
fit_entropy_formatted(data_out, alpha = NULL)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{alpha}{regularization parameter}
}
\value{
synthetic control weights
}
\description{
Fit entropy regularized synthetic controls
by solving the dual problem
}
