% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_multi.R
\name{fit_feff}
\alias{fit_feff}
\title{Get fixed effects from pre-treatment data for each level}
\usage{
fit_feff(X, trt, mask, force)
}
\arguments{
\item{X}{Matrix of outcomes}

\item{trt}{Vector of treatment status for each unit}

\item{mask}{Matrix of treatment statuses}

\item{force}{Fixed effects: 1="unit", 2="time", 3="two-way"}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Get fixed effects from pre-treatment data for each level
}
