% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_multi.R
\name{fit_gsynth_multi}
\alias{fit_gsynth_multi}
\title{Use gsynth to fit factor model with}
\usage{
fit_gsynth_multi(X, trt, r = 0, r.end = 5, force = 3, CV = 1)
}
\arguments{
\item{X}{Matrix of outcomes}

\item{trt}{Vector of treatment status for each unit}

\item{r}{Number of factors to use (or start with if CV==1)}

\item{r.end}{Max number of factors to consider if CV==1}

\item{CV}{Whether to do CV (0=no CV, 1=yes CV)}

\item{force=c(0, 1, 2, 3)}{Fixed effects (0=none, 1=unit, 2=time, 3=two-way)}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Use gsynth to fit factor model with
}
