% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_ipw}
\alias{fit_ipw}
\title{Fit IPW weights with a logit propensity score model}
\usage{
fit_ipw(outcomes, metadata, trt_unit = 1, alpha_w)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alpha_w}{regularization parameter for weights}
}
\value{
inverse of predicted propensity scores
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Fit IPW weights with a logit propensity score model
}
