% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_l1_entropy}
\alias{fit_l1_entropy}
\title{Fit entropy regularized synthetic controls with l1 error bounds}
\usage{
fit_l1_entropy(outcomes, metadata, trt_unit = 1, eps = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps}{width of error box around treated unit pre-treatment outcomes}
}
\value{
Weights for synthetic controls, control outcomes as matrix,
        and whether the unit is actually treated
}
\description{
Fit entropy regularized synthetic controls with l1 error bounds
}
