% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_l1_entropy_formatted}
\alias{fit_l1_entropy_formatted}
\title{Fit entropy regularized synthetic controls with l1 error bounds
by solving the dual problem}
\usage{
fit_l1_entropy_formatted(data_out, eps = NULL)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{eps}{width of error box around treated unit pre-treatment outcomes}
}
\value{
synthetic control weights
}
\description{
Fit entropy regularized synthetic controls with l1 error bounds
by solving the dual problem
}
