% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_l2_entropy_formatted}
\alias{fit_l2_entropy_formatted}
\title{Fit l2 entropy regularized synthetic controls
by solving the dual problem}
\usage{
fit_l2_entropy_formatted(data_out, eps = NULL)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{eps}{Bound on synthetic control differences}
}
\value{
synthetic control weights
}
\description{
Fit l2 entropy regularized synthetic controls
by solving the dual problem
}
