% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{fit_prog_causalimpact}
\alias{fit_prog_causalimpact}
\title{Fit a bayesian structural time series
to fit E[Y(0)|X]}
\usage{
fit_prog_causalimpact(X, y, trt, ...)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Model parameters}}
}
\description{
Fit a bayesian structural time series
to fit E[Y(0)|X]
}
