% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_prog_cits}
\alias{fit_prog_cits}
\title{Fit a Comparitive interupted time series
to fit E[Y(0)|X]}
\usage{
fit_prog_cits(X, y, trt, poly_order = 1, weights = NULL)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}

\item{poly_order}{Order of time trend polynomial to fit, default 1}

\item{weights}{Weights to use in WLS, default is no weights}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Fit a Comparitive interupted time series
to fit E[Y(0)|X]
}
