% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_prog_mcpanel}
\alias{fit_prog_mcpanel}
\title{Use Athey (2017) matrix completion panel data code}
\usage{
fit_prog_mcpanel(X, y, trt, unit_fixed = 1, time_fixed = 1)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}

\item{unit_fixed}{Whether to estimate unit fixed effects}

\item{time_fixed}{Whether to estimate time fixed effects}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Use Athey (2017) matrix completion panel data code
}
