% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_prog_reg}
\alias{fit_prog_reg}
\title{Use a separate regularized regression for each post period
to fit E[Y(0)|X]}
\usage{
fit_prog_reg(X, y, trt, alpha = 1, avg = FALSE)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}

\item{alpha}{Mixing between L1 and L2, default: 1 (LASSO)}

\item{avg}{Fit the average post-period rather than time periods separately}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Use a separate regularized regression for each post period
to fit E[Y(0)|X]
}
