% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_progsyn}
\alias{fit_progsyn}
\title{Fit E[Y(0)|X] and for each post-period and balance these}
\usage{
fit_progsyn(wide_data, synth_data, progfunc = c("EN", "RF", "GSYN",
  "CITS", "CausalImpact", "seq2seq"), weightfunc = c("SCM"),
  opts_out = NULL, opts_weights = NULL)
}
\arguments{
\item{wide_data}{Output of `format_data`}

\item{synth_data}{Output of `format_synth`}

\item{progfunc}{What function to use to impute control outcomes
EN=Elastic Net, RF=Random Forest, GSYN=gSynth
CITS=Comparative interrupted time series
CausalImpact=Bayesian structural time series with CausalImpact
seq2seq=Sequence to sequence learning with feedforward nets}

\item{weightfunc}{What function to use to fit weights
SCM=Vanilla Synthetic Controls}

\item{opts_out}{Optional options for fitting prognostic score}

\item{opts_weights}{Optional options for fitting synth weights}

\item{fit_weights}{Function to fit synth weights}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate (zero in this case)}
}
}
\description{
Fit E[Y(0)|X] and for each post-period and balance these
}
