% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_pscore_formatted}
\alias{fit_pscore_formatted}
\title{Fit IPW weights with a logit propensity score model}
\usage{
fit_pscore_formatted(data_out, alpha_w)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{alpha_w}{regularization parameter for weights}
}
\description{
Fit IPW weights with a logit propensity score model
}
