% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_residaug}
\alias{fit_residaug}
\title{Fit outcome model and balance residuals}
\usage{
fit_residaug(
  wide_data,
  synth_data,
  progfunc = c("GSYN", "MCP", "CITS", "CausalImpact"),
  weightfunc = c("SC", "ENT", "SVD", "NONE"),
  ...
)
}
\arguments{
\item{wide_data}{Output of `format_data`}

\item{synth_data}{Output of `format_synth`}

\item{progfunc}{What function to use to impute control outcomes
GSYN=gSynth, MCP=MCPanel,
CITS=Comparative interrupted time series
CausalImpact=Bayesian structural time series with CausalImpact}

\item{weightfunc}{What function to use to fit weights
SCM=Vanilla Synthetic Controls
NONE=No reweighting, just outcome model}

\item{...}{optional arguments for outcome model}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
}
}
\description{
Fit outcome model and balance residuals
}
