% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_residaug_formatted}
\alias{fit_residaug_formatted}
\title{Fit outcome model and balance residuals}
\usage{
fit_residaug_formatted(wide_data, synth_data, fit_progscore, fit_weights,
  opts_prog = NULL, opts_weights = NULL)
}
\arguments{
\item{wide_data}{Output of `format_data`}

\item{synth_data}{Output of `format_synth`}

\item{fit_progscore}{Function to fit prognostic score}

\item{fit_weights}{Function to fit synth weights}

\item{opts_weights}{Optional options for fitting synth weights}

\item{opts.gsyn}{Optional options for gsynth}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
}
}
\description{
Fit outcome model and balance residuals
}
