% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{fit_ridgeaug_cov_formatted}
\alias{fit_ridgeaug_cov_formatted}
\title{Ridge augmented weights with covariates}
\usage{
fit_ridgeaug_cov_formatted(ipw_format, syn_format, Z, lambda = NULL,
  scm = T)
}
\arguments{
\item{ipw_format}{Output of `format_ipw`}

\item{syn_format}{Output of `syn_format`}

\item{Z}{Matrix of covariates}

\item{lambda}{Ridge hyper-parameter, if NULL use CV}

\item{scm}{Include SCM or not}
}
\value{
inverse of predicted propensity scores
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Ridge augmented weights with covariates
}
