% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_screensyn_formatted}
\alias{fit_screensyn_formatted}
\title{Select covariates for E[Y(0)|X], then balance those}
\usage{
fit_screensyn_formatted(ipw_format, syn_format, screen_x, fit_weights,
  opts.prog = NULL, opts.weights = NULL)
}
\arguments{
\item{ipw_format}{Output of `format_ipw`}

\item{syn_format}{Output of `syn_format`}

\item{fit_weights}{Function to fit synth weights}

\item{opts.weights}{Optional options for fitting synth weights}

\item{fit_progscore}{Function to fit prognostic score}

\item{opts.screen}{Optional options for covariate screening}
}
\value{
inverse of predicted propensity scores
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Select covariates for E[Y(0)|X], then balance those
}
