% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{fit_synth}
\alias{fit_synth}
\title{Fit synthetic controls on outcomes, wrapper around fit_synth_formatted}
\usage{
fit_synth(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
Weights for synthetic controls, control outcomes as matrix,
        and whether the unit is actually treated
}
\description{
Fit synthetic controls on outcomes, wrapper around fit_synth_formatted
}
