% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{fit_synth_formatted}
\alias{fit_synth_formatted}
\title{Fit synthetic controls on outcomes after formatting data}
\usage{
fit_synth_formatted(synth_data, V = NULL)
}
\arguments{
\item{synth_data}{Panel data in format of Synth::dataprep}

\item{V}{Matrix to scale the obejctive by}
}
\value{
\itemize{
         \item{"weights"}{Synth weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
}
}
\description{
Fit synthetic controls on outcomes after formatting data
}
