% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_regression_multi.R
\name{fit_time_reg}
\alias{fit_time_reg}
\title{Fit a time regression}
\usage{
fit_time_reg(X, trt, n_leads, reg_param, lowlim = 0, uplim = 1, ...)
}
\arguments{
\item{X}{Matrix of outcomes}

\item{trt}{Vector of treatment status for each unit}

\item{n_leads}{How long past treatment effects should be estimated for}

\item{reg_param}{Regularization hyperparameter}

\item{lowlim}{Lower bound for coefs}

\item{uplim}{upper bound for coefs}

\item{...}{Extra optimization hyperparameters}
}
\value{
\itemize{
          \item{y0hat }{List of predicted outcome under control}
          \item{residuals }{List of residuals}
          \item{params }{Regression parameters}}
}
\description{
Fit a time regression
}
