% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_regression_multi.R
\name{fit_time_reg_qp_}
\alias{fit_time_reg_qp_}
\title{Fit a time regression}
\usage{
fit_time_reg_qp_(X, trt, n_leads, lowlim, uplim, reg_param, ...)
}
\arguments{
\item{X}{Matrix of outcomes}

\item{trt}{Vector of treatment status for each unit}

\item{n_leads}{How long past treatment effects should be estimated for}

\item{lowlim}{Lower bound for coefs}

\item{uplim}{upper bound for coefs}

\item{reg_param}{Regularization hyperparameter}

\item{...}{Extra optimization hyperparameters}
}
\value{
reg_weights Fitted regression weights
}
\description{
Fit a time regression
}
