% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{fit_uniform_formatted}
\alias{fit_uniform_formatted}
\title{Use difference in means}
\usage{
fit_uniform_formatted(data_out)
}
\arguments{
\item{data_out}{Output of format_ipw}
}
\value{
uniform weights
}
\description{
Use difference in means
}
