% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_multi.R
\name{fit_unit_feff}
\alias{fit_unit_feff}
\title{Get unit fixed effects from pre-treatment data for each level}
\usage{
fit_unit_feff(X, trt, mask)
}
\arguments{
\item{X}{Matrix of outcomes}

\item{trt}{Vector of treatment status for each unit}

\item{mask}{Matrix of treatment statuses}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Get unit fixed effects from pre-treatment data for each level
}
