% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_ipw}
\alias{format_ipw}
\title{Format "long" panel data into "wide" matrices to fit IPW}
\usage{
format_ipw(outcomes, metadata, outcome_col = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{outcome_col}{Column name which identifies outcomes,
if NULL then only one outcome is assumed}
}
\value{
List of pre and post period outcomes, and treatment indicator
}
\description{
Format "long" panel data into "wide" matrices to fit IPW
}
