% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R, R/format.R
\name{format_synth}
\alias{format_synth}
\alias{format_synth}
\title{Get the outcomes data into the correct form for Synth::synth}
\usage{
format_synth(outcomes, metadata, trt_unit = 1)

format_synth(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
List of data to use as an argument for Synth::synth,
        whether the unit was actually treated

List of data to use as an argument for Synth::synth,
        whether the unit was actually treated
}
\description{
Get the outcomes data into the correct form for Synth::synth

Get the outcomes data into the correct form for Synth::synth
}
