% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_synth_multi}
\alias{format_synth_multi}
\title{Get multiple outcomes data as matrices}
\usage{
format_synth_multi(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
List of data to use as an argument for Synth::synth,
        whether the unit was actually treated
}
\description{
Get multiple outcomes data as matrices
}
