% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{genoud_alpha}
\alias{genoud_alpha}
\title{Find the best regularization parameter with genetic optimization}
\usage{
genoud_alpha(outcomes, metadata, trt_unit, ...)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression)

Best alpha}
}
\description{
Find the best regularization parameter with genetic optimization
}
