% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancer.R
\name{get_balancer}
\alias{get_balancer}
\title{Find Balancing weights by solving the dual optimization problem}
\usage{
get_balancer(outcomes, metadata, trt_unit = 1, hyperparam, link = c("logit",
  "linear", "pos-linear"), regularizer = c(NULL, "l1", "l2", "ridge", "linf"),
  normalized = TRUE, outcome_col = NULL, cols = list(unit = "unit", time =
  "time", outcome = "outcome", treated = "treated"), opts = list())
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{hyperparam}{Regularization hyperparameter}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{normalized}{Whether to normalize the weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error tolerance}}}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Find Balancing weights by solving the dual optimization problem
}
