% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_dr}
\alias{get_dr}
\title{Fit a regularized outcome model and synthetic controls}
\usage{
get_dr(outcomes, metadata, trt_unit = 1, eps_w, alpha_o = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps_w}{List of bounds on synthetic control differences for each
outcome, if only one outcome type then a scalar}

\item{alpha_o}{regularization parameter for outcome model, if NULL use CV}
}
\value{
synthetic control weights,
        outcome regression parameters
        control outcomes
        boolean for treated
}
\description{
Fit a regularized outcome model and synthetic controls
}
