% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_ebal}
\alias{get_ebal}
\title{Fit entropy balancing weights}
\usage{
get_ebal(outcomes, metadata)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}
}
\value{
outcomes with imputed control values, weights, dual variables
}
\description{
Fit entropy balancing weights
}
