% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Fit entropy regularized synthetic controls on outcomes}
\usage{
get_entropy(outcomes, metadata, trt_unit = 1, alpha = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alpha}{regularization parameter}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit entropy regularized synthetic controls on outcomes
}
