% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{get_gsynaug}
\alias{get_gsynaug}
\title{Fit gsynth and balance residuals}
\usage{
get_gsynaug(outcomes, metadata, trt_unit = 1, weightfunc = c("SC", "ENT",
  "NONE"), opts.gsyn = NULL, opts.weights = NULL, outcome_col = NULL,
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{weightfunc}{What function to use to fit weights
SC=Vanilla Synthetic Controls, ENT=Maximum Entropy
NONE=No reweighting, just gsynth}

\item{opts.gsyn}{Optional options for gsynth}

\item{opts.weights}{Optional options for fitting synth weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit gsynth and balance residuals
}
