% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_outcomes.R
\name{get_index_syn}
\alias{get_index_syn}
\title{Fit synthetic controls for multiple outcomes together with an index}
\usage{
get_index_syn(outcomes, metadata, trt_unit = 1, alpha = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alpha}{Vector of size (n_outcomes - 1) or n_outcomes
which specifies the weights,
if no input then alpha = 1/n_units}

\item{syn_func}{Function to get synthetic controls}

\item{name}{Name of the synthetic control method}
}
\value{
both outcomes with additional synthetic controls added
        weights for synthetic controls
}
\description{
Fit synthetic controls for multiple outcomes together with an index
}
