% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_ipw}
\alias{get_ipw}
\title{Fit IPW weights with a logit propensity score model}
\usage{
get_ipw(outcomes, metadata, trt_unit = 1, alpha_w = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{alpha_w}{regularization parameter for weights}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit IPW weights with a logit propensity score model
}
