% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_outcomes.R
\name{get_joint_ent}
\alias{get_joint_ent}
\title{Fit entropy regularized synthetic controls for multiple outcomes with
the same weights across different outcomes}
\usage{
get_joint_ent(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
both outcomes with additional synthetic controls added
        weights for both synthetic controls
}
\description{
Fit entropy regularized synthetic controls for multiple outcomes with
the same weights across different outcomes
}
