% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_l1_entropy}
\alias{get_l1_entropy}
\title{Fit entropy regularized synthetic controls on outcomes}
\usage{
get_l1_entropy(outcomes, metadata, trt_unit = 1, eps = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps}{width of error box around treated unit pre-treatment outcomes}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit entropy regularized synthetic controls on outcomes
}
