% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{get_lambda_max}
\alias{get_lambda_max}
\title{Choose max lambda as largest eigenvalue of control X}
\usage{
get_lambda_max(X_c)
}
\arguments{
\item{X_c}{matrix of control lagged outcomes}
}
\value{
max lambda
}
\description{
Choose max lambda as largest eigenvalue of control X
}
