% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligible_donors.R
\name{get_nona_donors}
\alias{get_nona_donors}
\title{Get donors that don't have missing outcomes where treated units have outcomes}
\usage{
get_nona_donors(X, y, trt, n_lags, n_leads, time_cohort)
}
\description{
Get donors that don't have missing outcomes where treated units have outcomes
}
