% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{get_placebo_gaps}
\alias{get_placebo_gaps}
\title{Run augsynth letting each unit be the treated unit.  This
implements the abadie placebo test (e.g., from tobacco paper) but
with the augsynth package.}
\usage{
get_placebo_gaps(ascm, att = TRUE)
}
\arguments{
\item{tx.id}{Name/ID of treatment group (not stored in ascm)}
}
\value{
Dataframe With each row corresponding to a unit (including
  the original treated unit) and each column corresponding to a
  time point.  Each entry is the estimated impact for that unit at
  that time, after fitting augsynth on that unit as the tx unit and
  the other units as possible controls.
}
\description{
Run augsynth letting each unit be the treated unit.  This
implements the abadie placebo test (e.g., from tobacco paper) but
with the augsynth package.
}
