% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{get_progsyn}
\alias{get_progsyn}
\title{Fit synthetic controls on estimated outcomes under control}
\usage{
get_progsyn(outcomes, metadata, trt_unit = 1, progfunc = c("EN", "RF",
  "GSYN", "CITS", "CausalImpact"), weightfunc = c("SC", "ENT"),
  opts.prog = NULL, opts.weights = NULL, outcome_col = NULL,
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{progfunc}{What function to use to impute control outcomes
EN=Elastic Net, RF=Random Forest, GSYN=gSynth}

\item{weightfunc}{What function to use to fit weights
SC=Vanilla Synthetic Controls, ENT=Maximum Entropy}

\item{opts.prog}{Optional options for fitting prognostic score}

\item{opts.weights}{Optional options for fitting synth weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit synthetic controls on estimated outcomes under control
}
