% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{get_residaug}
\alias{get_residaug}
\title{Fit outcome model and balance residuals}
\usage{
get_residaug(outcomes, metadata, trt_unit = 1, progfunc = c("GSYN", "COMP"),
  weightfunc = c("SC", "ENT", "NONE"), opts.prog = NULL,
  opts.weights = NULL, outcome_col = NULL, cols = list(unit = "unit", time
  = "time", outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{progfunc}{What function to use to impute control outcomes
GSYN=gSynth, COMP=Matrix completion}

\item{weightfunc}{What function to use to fit weights
SC=Vanilla Synthetic Controls, ENT=Maximum Entropy
NONE=No reweighting, just gsynth}

\item{opts.prog}{Optional options for gsynth}

\item{opts.weights}{Optional options for fitting synth weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit outcome model and balance residuals
}
