% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{get_ridgeaug_cov}
\alias{get_ridgeaug_cov}
\title{Fit synthetic controls on outcomes}
\usage{
get_ridgeaug_cov(outcomes, metadata, covariates, trt_unit = 1,
  lambda = NULL, scm = T, cols = list(unit = "unit", time = "time",
  outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{lambda}{Ridge hyper-parameter, if NULL use CV}

\item{scm}{Include SCM or not}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{covaraites}{Dataframe of covariates}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit synthetic controls on outcomes
}
