% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_outcomes.R
\name{get_separate_ent}
\alias{get_separate_ent}
\title{Fit synthetic controls for multiple outcomes separately}
\usage{
get_separate_ent(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
both outcomes with additional synthetic controls added
        weights for both synthetic controls
}
\description{
Fit synthetic controls for multiple outcomes separately
}
