% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{get_synth}
\alias{get_synth}
\title{Fit synthetic controls on outcomes}
\usage{
get_synth(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit synthetic controls on outcomes
}
