% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_uniform}
\alias{get_uniform}
\title{Use difference in means}
\usage{
get_uniform(outcomes, metadata)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}
}
\value{
outcomes with imputed control values, weights, dual variables
}
\description{
Use difference in means
}
