% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{get_wlsaug}
\alias{get_wlsaug}
\title{Get weights then fit outcome model with weighted loss function}
\usage{
get_wlsaug(outcomes, metadata, trt_unit = 1, progfunc = c("GSYN", "COMP",
  "MCP", "CITS"), weightfunc = c("SC", "ENT", "SVD", "NONE"),
  opts.prog = NULL, opts.weights = NULL, outcome_col = NULL,
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{progfunc}{What function to use to impute control outcomes
GSYN=gSynth, COMP=softImpute, MCP=MCPanel}

\item{weightfunc}{What function to use to fit weights
SC=Vanilla Synthetic Controls, ENT=Maximum Entropy
SVD=SCM after dimension reduction,
NONE=No reweighting, just outcome model}

\item{opts.prog}{Optional options for gsynth}

\item{opts.weights}{Optional options for fitting synth weights}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Get weights then fit outcome model with weighted loss function
}
