% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_multisynth_absolute}
\alias{grad_multisynth_absolute}
\title{Gradient for Multiple synthetic controls and absolute time}
\usage{
grad_multisynth_absolute(theta, opts)
}
\description{
Gradient for Multiple synthetic controls and absolute time
}
