% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_multisynth_relative}
\alias{grad_multisynth_relative}
\title{Gradient for Multiple synthetic controls and relative time}
\usage{
grad_multisynth_relative(theta, opts)
}
\description{
Gradient for Multiple synthetic controls and relative time
}
