% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{impute_controls}
\alias{impute_controls}
\title{Impute the controls after fitting synth}
\usage{
impute_controls(outcomes, fit, trt_unit)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{fit}{Output of fit_synth}

\item{trt_unit}{Unit that is treated (target for regression)}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Impute the controls after fitting synth
}
