% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{impute_dr}
\alias{impute_dr}
\title{Impute the controls after fitting a dr estimator}
\usage{
impute_dr(outcomes, metadata, fit, trt_unit)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe with metadata, in particular a t_int column}

\item{fit}{Output of fit_dr}

\item{trt_unit}{Unit that is treated (target for regression)}
}
\value{
outcomes with additional synthetic control added,
        synth weights
        outcome regression weights
}
\description{
Impute the controls after fitting a dr estimator
}
