% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{impute_gsynaug}
\alias{impute_gsynaug}
\title{Impute the controls after fitting gynsth and reweighting residuals}
\usage{
impute_gsynaug(outcomes, metadata, fit, trt_unit)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe with metadata, in particular a t_int column}

\item{fit}{Output of fit_gsynaug_formatted}
}
\value{
outcomes with additional synthetic control added,
        synth weights
        outcome regression weights
}
\description{
Impute the controls after fitting gynsth and reweighting residuals
}
