% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{jackknife}
\alias{jackknife}
\title{Compute standard errors using the jackknife}
\usage{
jackknife(multisynth, relative = NULL)
}
\arguments{
\item{multisynth}{fitted multisynth object}

\item{relative}{Whether to compute effects according to relative time}
}
\description{
Compute standard errors using the jackknife
}
