% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{jackknife_se_multi}
\alias{jackknife_se_multi}
\title{Compute standard errors using the jackknife}
\usage{
jackknife_se_multi(multisynth, relative = NULL)
}
\arguments{
\item{multisynth}{fitted multisynth object}

\item{relative}{Whether to compute effects according to relative time}
}
\description{
Compute standard errors using the jackknife
}
