% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{jackknife_se_multiout}
\alias{jackknife_se_multiout}
\title{Estimate standard errors for multi outcome ascm with jackknife}
\usage{
jackknife_se_multiout(ascm)
}
\arguments{
\item{ascm}{Fitted augsynth object}
}
\description{
Estimate standard errors for multi outcome ascm with jackknife
}
