% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{jackknife_se_ridgeaug}
\alias{jackknife_se_ridgeaug}
\title{Estimate standard errors with the jackknife
Do this for ridge-augmented synth}
\usage{
jackknife_se_ridgeaug(wide_data, synth_data, Z = NULL, lambda = NULL,
  ridge = T, scm = T, fixedeff = F)
}
\arguments{
\item{wide_data}{Data formatted from format_data}

\item{synth_data}{Data formatted from foramt_synth}

\item{Z}{Matrix of auxiliary covariates}

\item{lambda}{Ridge hyper-parameter, if NULL use CV}

\item{ridge}{Include ridge or not}

\item{scm}{Include SCM or not}

\item{fixedeff}{Take out fixed effects or not}
}
\value{
att estimates, test statistics, p-values
}
\description{
Estimate standard errors with the jackknife
Do this for ridge-augmented synth
}
