% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{lasso_screen}
\alias{lasso_screen}
\title{Screen covariates for the outcome process}
\usage{
lasso_screen(ipw_format, syn_format, alpha = 1, type = "sep")
}
\arguments{
\item{ipw_format}{Output of `format_ipw`}

\item{syn_format}{Output of `syn_format`}

\item{alpha}{Elastic Net parameter}

\item{type}{How to fit outcome model(s)
\itemize{
 \item{sep }{Separate outcome models}
 \item{avg }{Average responses into 1 outcome}
 \item{multi }{Use multi response regression in glmnet}}}
}
\value{
\itemize{
          \item{selX }{Selected covariates}
          \item{params }{Regression parameters}}
}
\description{
Screen covariates for the outcome process
}
