% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{lexical}
\alias{lexical}
\title{Finds the lowest feasible tolerance in each group, holding
the previous groups fixed}
\usage{
lexical(outcomes, metadata, grp_order, outcome_col, trt_unit = 1, by = 0.1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{grp_order}{Lexical ordering of groups}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{by}{Step size for tolerances to dry, default: 0.1 * magnitude}
}
\value{
List with lowest tolerances lexically
}
\description{
Finds the lowest feasible tolerance in each group, holding
the previous groups fixed
}
