% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{lexical_time}
\alias{lexical_time}
\title{Finds the lowest feasible tolerance in each time period from
most recent to oldest, holding the previous times fixed}
\usage{
lexical_time(outcomes, metadata, trt_unit = 1, by = 0.1, cols = list(unit
  = "unit", time = "time", outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{by}{Step size for tolerances to dry, default: 0.1 * magnitude}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
List with lowest tolerances lexically
}
\description{
Finds the lowest feasible tolerance in each time period from
most recent to oldest, holding the previous times fixed
}
