% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{logsumexp_grad}
\alias{logsumexp_grad}
\title{Compute numerically stable logsumexp gradient with natural param eta
and data x}
\usage{
logsumexp_grad(eta, x)
}
\description{
Compute numerically stable logsumexp gradient with natural param eta
and data x
}
