% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth_qp.R
\name{make_constraint_mats}
\alias{make_constraint_mats}
\title{Create constraint matrices for multisynth QP}
\usage{
make_constraint_mats(trt, grps, n_leads, n_lags, Xc, d)
}
\arguments{
\item{trt}{Vector of treatment levels/times}

\item{grps}{Treatment times}

\item{n_leads}{Number of time periods after treatment to impute control values.}

\item{n_lags}{Number of pre-treatment periods to balance}

\item{Xc}{List of outcomes for possible comparison units}

\item{d}{Max number of lagged outcomes}
}
\value{
\itemize{
         \item{"Amat"}{Linear constraint matrix}
         \item{"lvec"}{Lower bounds for linear constraints}
         \item{"uvec"}{Upper bounds for linear constraints}
        }
}
\description{
Create constraint matrices for multisynth QP
}
