% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth_qp.R
\name{make_qvec}
\alias{make_qvec}
\title{Make the vector in the QP}
\usage{
make_qvec(Xc, x_t, nu, n_lags, d)
}
\arguments{
\item{Xc}{List of outcomes for possible comparison units}

\item{x_t}{List of outcomes for treated units}

\item{nu}{Hyperparameter between global and individual balance}

\item{n_lags}{Number of lags to balance}

\item{d}{Largest number of pre-intervention time periods}
}
\description{
Make the vector in the QP
}
