% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth.R
\name{map_to_param}
\alias{map_to_param}
\title{Map string choices to the proper parameters for the balancer sub-functions}
\usage{
map_to_param(X, link = c("logit", "linear", "pos-linear", "pos-enet",
  "posenet"), regularizer = c("nuc", "ridge"), alpha = 1)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{alpha}{Elastic net parameter \eqn{\frac{1-\alpha}{2}\|\beta\|_2^2 + \alpha\|\beta\|_1}, defaults to 1}

\item{type}{Find balancing weights for ATT, subgroup ATTs,
subgroup ATTs with multilevel p-score, multilevel observational studies,
ATT with missing outcomes, and heterogeneous effects}
}
\value{
Parameters for balancer
}
\description{
Map string choices to the proper parameters for the balancer sub-functions
}
