% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth.R
\name{multisynth_}
\alias{multisynth_}
\title{Internal function to fit synth with staggered adoption}
\usage{
multisynth_(X, trt, mask, relative, gap, link = c("logit", "linear",
  "pos-linear", "pos-enet", "posenet"), regularizer = c("nuc", "ridge"),
  lambda = NULL, nlambda = 20, lambda.min.ratio = 0.001, alpha = 1,
  opts = list())
}
\arguments{
\item{X}{Matrix of pre-final intervention outcomes}

\item{trt}{Vector of treatment levels/times}

\item{relative}{Whether to re-index time according to treatment date, default T}

\item{gap}{Number of time periods after treatment to impute control values.
For units treated at time T_j, all units treated after T_j + gap
will be used as control values. If larger than the number of periods,
only never never treated units (pure controls) will be used as comparison units}

\item{link}{Link function/dispersion function. Default is logit, for internal use only}

\item{regularizer}{Form of pooling. "multi_ridge" pools parameters towards the group mean.
"ridge" pools parameters towards the group mean and
"nuc" imposes a low-rank restriction on the parameters.}

\item{lambda}{Regularization hyper-parameter. If NULL then solutions will be computed for a
range of lambdas equally spaced on the log scale.}

\item{nlambda}{Number of values of lambda to consider if lambda is NULL}

\item{lambda.min.ratio}{Ratio between largest and smallest lambda values.}

\item{alpha}{Hyper-parameter that controls trade-off between overall and individual balance.
Larger values of alpha place more emphasis on individual balance.
Regularization is
  (1-alpha) * lambda ||global|| + alpha * lambda ||individual||}

\item{opts}{Additional options for optimization}
}
\description{
Internal function to fit synth with staggered adoption
}
