% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth.R
\name{multisynth_absolute_}
\alias{multisynth_absolute_}
\title{Internal function to fit synth with staggered adoption according to absolute time}
\usage{
multisynth_absolute_(X, trt, mask, weightfunc, weightfunc_ptr, proxfunc,
  balancefunc, lambda = NULL, nlambda = 20, lambda.min.ratio = 0.001,
  opts = list(), prox_opts = list())
}
\description{
Internal function to fit synth with staggered adoption according to absolute time
}
