% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth_qp.R
\name{multisynth_absolute_qp}
\alias{multisynth_absolute_qp}
\title{Internal function to fit synth with staggered adoption according to absolute time
#' @importMethodsFrom Matrix %*%}
\usage{
multisynth_absolute_qp(X, trt, mask, gap, alpha, lambda, ...)
}
\description{
Internal function to fit synth with staggered adoption according to absolute time
#' @importMethodsFrom Matrix %*%
}
