% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{multisynth_formatted}
\alias{multisynth_formatted}
\title{Internal funciton to fit staggered synth with formatted data}
\usage{
multisynth_formatted(
  wide,
  relative = T,
  n_leads,
  n_lags,
  nu,
  lambda,
  force,
  n_factors,
  scm,
  time_cohort,
  time_w,
  lambda_t,
  fit_resids,
  eps_abs,
  eps_rel,
  verbose,
  ...
)
}
\arguments{
\item{wide}{List containing data elements}

\item{relative}{Whether to compute balance by relative time}

\item{n_leads}{How long past treatment effects should be estimated for}

\item{n_lags}{Number of pre-treatment periods to balance, default is to balance all periods}

\item{nu}{Fraction of balance for individual balance}

\item{lambda}{Regularization hyperparameter, default = 0}

\item{force}{c(0,1,2,3) what type of fixed effects to include}

\item{n_factors}{Number of factors for interactive fixed effects, default does CV}

\item{scm}{Whether to fit scm weights}

\item{time_cohort}{Whether to average synthetic controls into time cohorts}

\item{time_w}{Whether to fit time weights}

\item{fit_resids}{Whether to fit SCM on the residuals or not}

\item{eps_abs}{Absolute error tolerance for osqp}

\item{eps_rel}{Relative error tolerance for osqp}

\item{verbose}{Whether to print logs for osqp}

\item{...}{Extra arguments}
}
\value{
multisynth object
}
\description{
Internal funciton to fit staggered synth with formatted data
}
