% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{multisynth_old}
\alias{multisynth_old}
\title{Fit staggered synth
DEPRECATED}
\usage{
multisynth_old(form, unit, time, data, relative = T, gap = NULL,
  lambda = NULL, alpha = 0.5, force = "two-way", n_factors = NULL,
  opts_weights = NULL)
}
\arguments{
\item{form}{outcome ~ treatment | auxillary covariates}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{data}{Panel data as dataframe}

\item{relative}{Whether to compute balance by relative time}

\item{gap}{How long past treatment effects should be estimated for}

\item{lambda}{Regularization hyperparameter}

\item{alpha}{Fraction of balance for individual balance}

\item{force}{Include "none", "unit", "time", "two-way" fixed effects. Default: "two-way"}

\item{n_factors}{Number of factors for interactive fixed effects}

\item{opts_weights}{Optional options for fitting synth weights}
}
\value{
augsynth object that contains:
        \itemize{
         \item{"weights"}{weights}
         \item{"data"}{Panel data as matrices}
        }
}
\description{
Fit staggered synth
DEPRECATED
}
