% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth_qp.R
\name{multisynth_qp}
\alias{multisynth_qp}
\title{Internal function to fit synth with staggered adoption with a QP solver}
\usage{
multisynth_qp(X, trt, mask, n_leads = NULL, n_lags = NULL,
  relative = T, nu = 0, lambda = 0, time_cohort = FALSE,
  verbose = FALSE, eps_rel = 1e-04, eps_abs = 1e-04)
}
\arguments{
\item{X}{Matrix of pre-final intervention outcomes, or list of such matrices after transformations}

\item{trt}{Vector of treatment levels/times}

\item{mask}{Matrix with indicators for observed pre-intervention times for each treatment group}

\item{n_leads}{Number of time periods after treatment to impute control values.
For units treated at time T_j, all units treated after T_j + n_leads
will be used as control values. If larger than the number of periods,
only never never treated units (pure controls) will be used as comparison units}

\item{n_lags}{Number of pre-treatment periods to balance, default is to balance all periods}

\item{relative}{Whether to re-index time according to treatment date, default T}

\item{nu}{Hyper-parameter that controls trade-off between overall and individual balance.
Larger values of nu place more emphasis on individual balance.
Balance measure is
  nu ||global|| + (1-nu) ||individual||
Default: 0}

\item{lambda}{Regularization hyper-parameter. Default, 0}

\item{time_cohort}{Whether to average synthetic controls into time cohorts}

\item{verbose}{Whether to print logs for osqp}

\item{eps_rel}{Relative error tolerance for osqp}

\item{eps_abs}{Absolute error tolerance for osqp}
}
\value{
\itemize{
         \item{"weights"}{Matrix of unit weights}
         \item{"imbalance"}{Matrix of overall and group specific imbalance}
         \item{"global_l2"}{Imbalance overall}
         \item{"ind_l2"}{Matrix of imbalance for each group}
        }
}
\description{
Internal function to fit synth with staggered adoption with a QP solver
}
