% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth.R
\name{multisynth_relative_}
\alias{multisynth_relative_}
\title{Internal function to fit synth with staggered adoption according to time relative to treatment}
\usage{
multisynth_relative_(X, trt, mask, gap, weightfunc, weightfunc_ptr,
  proxfunc, balancefunc, lambda = NULL, nlambda = 20,
  lambda.min.ratio = 0.001, opts = list(), prox_opts = list())
}
\description{
Internal function to fit synth with staggered adoption according to time relative to treatment
}
