% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_inf}
\alias{permutation_inf}
\title{Generate formatted outputs for statistical inference using permutation inference}
\usage{
permutation_inf(augsynth, inf_type)
}
\arguments{
\item{augsynth}{An augsynth object.}
}
\description{
Generate formatted outputs for statistical inference using permutation inference
}
