% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{placebo_distribution}
\alias{placebo_distribution}
\title{Get placebo distribution}
\usage{
placebo_distribution(augsynth)
}
\arguments{
\item{augsynth}{Augsynth object or summery object with permutation inference of some sort.}
}
\value{
Data frame holding the placebo distribution, one row per placebo unit and time point.
}
\description{
Get placebo distribution
}
