% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{plot.summary.augsynth}
\alias{plot.summary.augsynth}
\title{Plot results from summarized augsynth}
\usage{
\method{plot}{summary.augsynth}(x, plot_type = "estimate", inf = TRUE, ...)
}
\arguments{
\item{x}{A `summary.augsynth` object.}

\item{plot_type}{Type of plot to return.}

\item{inf}{Logical; if `TRUE`, include uncertainty information when
available.}

\item{...}{Optional additional arguments.}
}
\description{
Make a variety of plots, depending on `plot_type`. The default is
to plot treatment effect estimates with associated uncertainty (if
present).
}
\details{
If `inf` is true, will plot the inference results stored in the
summary object. Thus, the displayed confidence intervals and
p-values correspond to the `inf_type` used when `summary()` was
called, and are not recomputed at plot time.
}
