% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{plot.summary.augsynth_multiout}
\alias{plot.summary.augsynth_multiout}
\title{Plot function for summary function for augsynth}
\usage{
\method{plot}{summary.augsynth_multiout}(x, inf = F, plt_avg = F, ...)
}
\arguments{
\item{x}{summary.augsynth_multiout object}

\item{inf}{Boolean, whether to plot uncertainty intervals, default TRUE}

\item{plt_avg}{Boolean, whether to plot the average of the outcomes, default FALSE}

\item{...}{Optional arguments}
}
\description{
Plot function for summary function for augsynth
}
