% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_att}
\alias{plot_att}
\title{Plot the estimate of the att}
\usage{
plot_att(outcomes, metadata, trt_unit = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes}

\item{metadata}{Dataframe with info about the simulation}

\item{trt_unit}{Unit to count as treated, defaults to using
the treated column}
}
\description{
Plot the estimate of the att
}
