% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_outcomes}
\alias{plot_outcomes}
\title{Plot the outcomes from a study/simulation}
\usage{
plot_outcomes(outcomes, metadata, trt_unit = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes}

\item{metadata}{Dataframe with info about the simulation}

\item{trt_unit}{Unit to count as treated, defaults to using
the treated column}
}
\description{
Plot the outcomes from a study/simulation
}
