% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{predict.augsynth}
\alias{predict.augsynth}
\title{Get prediction of ATT or average outcome under control}
\usage{
\method{predict}{augsynth}(augsynth, type = "outcome")
}
\arguments{
\item{augsynth}{augsynth object}

\item{type}{Either "outcome" for average outcome under control, or "att" for ATT}
}
\value{
Vector of predicted post-treatment control averages
}
\description{
Get prediction of ATT or average outcome under control
}
