% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{predict.augsynth_multiout}
\alias{predict.augsynth_multiout}
\title{Get prediction of ATT or average outcome under control}
\usage{
\method{predict}{augsynth_multiout}(object, ...)
}
\arguments{
\item{object}{augsynth_multiout object}

\item{...}{Optional arguments, including \describe{\item{"att"}{Whether to return the ATT or average outcome under control}}}
}
\value{
Vector of predicted post-treatment control averages
}
\description{
Get prediction of ATT or average outcome under control
}
