% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess covariates}
\usage{
preprocess(X, trt, ipw_weights, type, link)
}
\arguments{
\item{X}{n x d matrix of covariates}

\item{trt}{Vector of treatment status indicators}

\item{ipw_weights}{Optional ipw weights to measure dispersion against}

\item{type}{Find balancing weights for ATT, subgroup ATTs,
subgroup ATTs with multilevel p-score, multilevel observational studies,
ATT with missing outcomes, and heterogeneous effects#'}

\item{link}{Link function for weights}

\item{normalized}{Whether to normalize the weights}
}
\value{
Processed covariate matrix
}
\description{
Preprocess covariates
}
