% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth_class.R
\name{print.augsynth}
\alias{print.augsynth}
\title{Print function for augsynth}
\usage{
\method{print}{augsynth}(x, ...)
}
\arguments{
\item{x}{augsynth object}

\item{...}{Optional arguments}
}
\description{
Print function for augsynth
}
