% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{print.multisynth}
\alias{print.multisynth}
\title{Print function for multisynth}
\usage{
\method{print}{multisynth}(x, att_weight = NULL, ...)
}
\arguments{
\item{x}{multisynth object}

\item{att_weight}{Weights to place on individual units/groups when averaging}

\item{...}{Optional arguments}
}
\description{
Print function for multisynth
}
