% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{print.summary.augsynth}
\alias{print.summary.augsynth}
\title{Print function for summary function for augsynth}
\usage{
\method{print}{summary.augsynth}(x, ...)
}
\arguments{
\item{x}{summary object}

\item{...}{Optional arguments}
}
\description{
Print function for summary function for augsynth
}
