% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proj_multilevel_weighted_ridge_nuc}
\alias{proj_multilevel_weighted_ridge_nuc}
\title{Squared L2 Prox for global parameters, nuclear norm projection for local parameters}
\usage{
proj_multilevel_weighted_ridge_nuc(x, lam, opts)
}
\arguments{
\item{x}{Input matrix (contains global and local parameters}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["alpha"] holds the ratio between global and local balance}
}
\value{
L2 squared prox values
}
\description{
Squared L2 Prox for global parameters, nuclear norm projection for local parameters
}
