% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{proj_nuc}
\alias{proj_nuc}
\title{Nuclear norm projection ||x||_* <= lam}
\usage{
proj_nuc(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Constraint on nuclear norm}

\item{opts}{List of options (opts["lam"] holds the other scaling}
}
\value{
Singular value soft thresholded X
}
\description{
Nuclear norm projection ||x||_* <= lam
}
