% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{prox_group}
\alias{prox_group}
\title{prox operator for group LASSO (generalization of prox_l2)}
\usage{
prox_group(x, lams, groups)
}
\arguments{
\item{x}{Vector to prox}

\item{lams}{List of prox scalings for each group}

\item{list}{of group indices}
}
\description{
prox operator for group LASSO (generalization of prox_l2)
}
