% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_multilevel_weighted_ridge_nuc}
\alias{prox_multilevel_weighted_ridge_nuc}
\title{Prox for weighted ridge on global params and nuclear norm on individual params}
\usage{
prox_multilevel_weighted_ridge_nuc(x, lam, opts)
}
\arguments{
\item{x}{Input matrix (contains global and local parameters}

\item{lam}{Prox scaling factor}

\item{opts}{List of options (opts["alpha"] holds the ratio between global and local balance}
}
\value{
L2 squared prox values
}
\description{
Prox for weighted ridge on global params and nuclear norm on individual params
}
