% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_weighted_l2_sq}
\alias{prox_weighted_l2_sq}
\title{Weighted squared L2 prox}
\usage{
prox_weighted_l2_sq(x, lam, opts)
}
\arguments{
\item{x}{Input matrix}

\item{lam}{Prox scaling factor}

\item{opts}{List of options. opts["lam"] holds the other scaling, opts["w"] holds unit specific weights}
}
\value{
Column soft thresholded X
}
\description{
Weighted squared L2 prox
}
