% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{recent_group}
\alias{recent_group}
\title{Lexically minimizes the imbalance in two groups, recent and old}
\usage{
recent_group(outcomes, metadata, t_past, trt_unit = 1, by = 0.1,
  maxep = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{t_past}{Number of recent time periods to balance especially}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{by}{Step size for tolerances to dry, default: 0.1 * magnitude}

\item{maxep}{Maximum tolerance to consider}
}
\value{
Fitted SC with lexically minimized imbalance in time groups
}
\description{
Lexically minimizes the imbalance in two groups, recent and old
}
