% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{RMSPE}
\alias{RMSPE}
\title{RMSPE for treated unit}
\usage{
RMSPE(augsynth)
}
\arguments{
\item{augsynth}{Augsynth object}
}
\value{
RMSPE (Root mean squared predictive error) for the treated unit in pre-treatment era
}
\description{
RMSPE for treated unit
}
