% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{sep_lasso}
\alias{sep_lasso}
\title{Finds the lowest feasible tolerance in units of standard deviation for
all time periods and covariates}
\usage{
sep_lasso(outcomes, metadata, trt_unit = 1, by = 0.1, scale = TRUE,
  maxep = 4, cols = list(unit = "unit", time = "time", outcome = "outcome",
  treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{by}{Step size for tolerances to dry, default: 0.1 * magnitude}

\item{scale}{Scale imbalances by standard deviations, default: True}

\item{maxep}{Maximum imbalance to consider in units of sd, default: 4
time variable, outcome, and treated indicator}

\item{cols}{Column names corresponding to the units,}
}
\value{
max ent SC fit with lowest imbalance tolerance
}
\description{
Finds the lowest feasible tolerance in units of standard deviation for
all time periods and covariates
}
