% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_outcomes.R
\name{sim_factor_model}
\alias{sim_factor_model}
\title{Generate data from a factor model as in ADH 2010 (5) and Firpo 2017 (21)}
\usage{
sim_factor_model(n_units, t_total, t_int, d, lambda, corr = 0,
  fac_size = 10, effect_type = "constant", sim_num = 1, frac_same = 0.3)
}
\arguments{
\item{n_units}{Number of units (first is always treated}

\item{t_total}{Total number of time steps}

\item{t_int}{Time of intervention}

\item{d}{Dimension of predictors}

\item{lambda}{Effect size}

\item{corr}{Correlation between outcomes}

\item{fac_size}{Dimension of hidden factors}

\item{effect_type}{Type of effect}

\item{sim_num}{Simulation number, defaults to 1}

\item{frac_same}{Fraction of units that are similar to the treated unit}
}
\value{
\itemize{
          \item{outcomes}{Tidy dataframe of outcomes}
          \item{metadata}{Metadata about the simulation}
        }
}
\description{
Generate data from a factor model as in ADH 2010 (5) and Firpo 2017 (21)
}
