% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{standard_error}
\alias{standard_error}
\title{Estimate the variance of the SC estimate}
\usage{
standard_error(outcomes, metadata, trt_unit, use_weights = FALSE,
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Treated unit}

\item{use_weights}{Whether to use weights in se estimate, default: FALSE}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{fitfunc}{Partially applied fitting function which takes in
the number of the treated unit}

\item{units}{Numbers of the units to include when estimating the se}
}
\value{
att estimates, test statistics, p-values
}
\description{
Estimate the variance of the SC estimate
}
