% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{standard_error_}
\alias{standard_error_}
\title{Internal function to estimate the variance of the SC estimate}
\usage{
standard_error_(metadata, fitfunc, units, trt_unit, posttimes, weights = NULL)
}
\arguments{
\item{metadata}{with treatment time}

\item{fitfunc}{Partially applied fitting function which takes in
the number of the treated unit}

\item{units}{Numbers of the control units to include when estimating the se}

\item{trt_unit}{Treated unit}

\item{posttimes}{Vector of times in post-period}

\item{weights}{Weights to use in SC estimate, default: NULL (Doudchenko-Imbens 2017)}
}
\value{
att estimates, test statistics, p-values
}
\description{
Internal function to estimate the variance of the SC estimate
}
