% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{is_summary_augsynth}
\alias{is_summary_augsynth}
\title{Methods for accessing details of summary.augsynth object}
\usage{
is_summary_augsynth(x)
}
\arguments{
\item{x}{summary.augsynth result object}
}
\value{
is_summary_augsynth: TRUE if object is a augsynth object.
}
\description{
Methods for accessing details of summary.augsynth object
}
