% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{summary.multisynth}
\alias{summary.multisynth}
\title{Summary function for multisynth}
\usage{
\method{summary}{multisynth}(object, inf_type = "bootstrap", att_weight = NULL, ...)
}
\arguments{
\item{object}{multisynth object}

\item{inf_type}{Type of inference to perform:
 \describe{
   \item{bootstrap}{Wild bootstrap, the default option}
   \item{jackknife}{Jackknife}
}}

\item{att_weight}{Weights to place on individual units/groups when averaging}

\item{...}{Optional arguments}
}
\value{
summary.multisynth object that contains:
         \item{"att"}{Dataframe with ATT estimates, standard errors for each treated unit}
         \item{"global_l2"}{L2 imbalance for the pooled synthetic control}
         \item{"scaled_global_l2"}{L2 imbalance for the pooled synthetic control, scaled by the imbalance for unitform weights}
         \item{"ind_l2"}{Average L2 imbalance for the individual synthetic controls}
         \item{"scaled_ind_l2"}{Average L2 imbalance for the individual synthetic controls, scaled by the imbalance for unitform weights}
        \item{"n_leads", "n_lags"}{Number of post treatment outcomes (leads) and pre-treatment outcomes (lags) to include in the analysis}
}
\description{
Summary function for multisynth
}
