% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{svyglm_se_}
\alias{svyglm_se_}
\title{Compute standard errors from survey glm}
\usage{
svyglm_se_(X = NULL, y, trt, weights, pred_int)
}
\arguments{
\item{X}{Matrix of pre-period outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Treatment indicator}

\item{weights}{Balancing weights}

\item{pred_int}{Whether to add outcome control variance to make pred interval}
}
\description{
Compute standard errors from survey glm
}
