% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{svyglm_se_synth}
\alias{svyglm_se_synth}
\title{svyglm standard errors}
\usage{
svyglm_se_synth(outcomes, metadata, trt_unit = 1, pred_int = f,
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Treated unit}

\item{pred_int}{Whether to add outcome control variance to make pred interval}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
svyglm standard errors
}
