% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{swap_treat_unit}
\alias{swap_treat_unit}
\title{Take inner data and change which unit is marked as 'treated'}
\usage{
swap_treat_unit(wide_data, Z, i)
}
\description{
Take inner data and change which unit is marked as 'treated'
}
