% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{n_unit}
\alias{n_unit}
\alias{n_time}
\alias{n_treated}
\title{Number of time points in fit data}
\usage{
n_unit(x, ...)

n_time(x, ...)

n_treated(x, ...)
}
\arguments{
\item{x}{Object to compute on}

\item{...}{Optional arguments}
}
\value{
Single number (of unique units).

Single number (of unique time points).

Single number (of number of treated units).
}
\description{
Number of time points in fit data

Number of treated units in fit data
}
