% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{synth_qp}
\alias{synth_qp}
\title{Solve the synth QP directly}
\usage{
synth_qp(X1, X0, V)
}
\arguments{
\item{X1}{Target vector}

\item{X0}{Matrix of control outcomes}

\item{V}{Scaling matrix}
}
\description{
Solve the synth QP directly
}
