% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{time_jackknife_plus}
\alias{time_jackknife_plus}
\title{Jackknife+ algorithm over time}
\usage{
time_jackknife_plus(ascm, alpha = 0.05, conservative = F)
}
\arguments{
\item{ascm}{Fitted `augsynth` object}

\item{alpha}{Confidence level}

\item{conservative}{Whether to use the conservative jackknife+ procedure}
}
\value{
List that contains:
         \item{"att"}{Vector of ATT estimates}
         \item{"heldout_att"}{Vector of ATT estimates with the time period held out}
         \item{"se"}{Standard error, always NA but returned for compatibility}
         \item{"lb"}{Lower bound of 1 - alpha confidence interval}
         \item{"ub"}{Upper bound of 1 - alpha confidence interval}
         \item{"alpha"}{Level of confidence interval}
}
\description{
Jackknife+ algorithm over time
}
