% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth_class.R
\name{treated_table}
\alias{treated_table}
\title{Return a summary data frame for the treated unit}
\usage{
treated_table(augsynth)
}
\arguments{
\item{augsynth}{Augsynth object of interest}
}
\value{
Dataframe of information about the treated unit, one row
  per time point.  This includes the measured outcome, predicted
  outcome from the synthetic unit, the average of all donor units
  (as reference, called `raw_average`), and the estimated impact
  (`ATT`), and the r-statistic (ATT divided by RMSPE).
}
\description{
Return a summary data frame for the treated unit
}
\seealso{
[donor_table()]
}
