% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{weighted_test}
\alias{weighted_test}
\title{Estimate p-value with non-uniform probabilities of treatment with
SCM weights}
\usage{
weighted_test(metadata, fitfunc, units, probs, trt_unit, pretimes, posttimes,
  statfuncs)
}
\arguments{
\item{metadata}{with treatment time}

\item{fitfunc}{Partially applied fitting function which takes in
the number of the treated unit}

\item{units}{Numbers of the units to permute treatment around}

\item{probs}{Propensity scores}

\item{trt_unit}{Treated unit}

\item{pretimes}{Vector of times in pre-period}

\item{posttimes}{Vector of times in post-period}

\item{statfuncs}{Function to compute test stats}
}
\value{
att estimates, test statistics, p-values
}
\description{
Estimate p-value with non-uniform probabilities of treatment with
SCM weights
}
