% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{weighted_test_bal}
\alias{weighted_test_bal}
\title{Get the weighted permutation distribution of SC estimate test statistics
estimating p-scores with logit-link synth}
\usage{
weighted_test_bal(outcomes, metadata, trt_unit, hyperparam, link = c("logit",
  "linear", "pos-linear"), regularizer = c(NULL, "l1", "l2", "ridge", "linf"),
  normalized = TRUE, statfuncs = c(rmse_ratio, mean_abs, abs_tstat),
  cols = list(unit = "unit", time = "time", outcome = "outcome", treated =
  "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{statfuncs}{Function to compute test stats}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{n_sim}{Number of montecarlo samples}
}
\value{
att estimates, test statistics, p-values
}
\description{
Get the weighted permutation distribution of SC estimate test statistics
estimating p-scores with logit-link synth
}
