% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{wls_se_bal}
\alias{wls_se_bal}
\title{Bootstrap standard errors for estimated counterfactual with synth}
\usage{
wls_se_bal(outcomes, metadata, hyperparam, trt_unit = 1, link = c("logit",
  "linear", "pos-linear"), regularizer = c(NULL, "l1", "l2", "ridge", "linf"),
  normalized = TRUE, cols = list(unit = "unit", time = "time", outcome =
  "outcome", treated = "treated"), opts = list())
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{hyperparam}{Regularization hyperparameter}

\item{trt_unit}{Treated unit}

\item{link}{Link function for weights}

\item{regularizer}{Dual of balance criterion}

\item{normalized}{Whether to normalize the weights}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{opts}{Optimization options
\itemize{
  \item{MAX_ITERS }{Maximum number of iterations to run}
  \item{EPS }{Error tolerance}}}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Bootstrap standard errors for estimated counterfactual with synth
}
