% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{wls_se_ridgeaug}
\alias{wls_se_ridgeaug}
\title{Weighted Least Squares Standard Errors}
\usage{
wls_se_ridgeaug(outcomes, metadata, trt_unit = 1, pred_int = F, hc = 0,
  lambda = NULL, scm = T, cols = list(unit = "unit", time = "time",
  outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Treated unit}

\item{pred_int}{Whether to add outcome control variance to make pred interval}

\item{hc}{Type of HC variance estimate (-1 for homoskedastic)}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Weighted Least Squares Standard Errors
}
