% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{wpermtest_sc}
\alias{wpermtest_sc}
\title{Get the weighted permutation distribution of SC estimate test statistics
estimating p-scores with logit-link synth}
\usage{
wpermtest_sc(outcomes, metadata, trt_unit, statfuncs = c(rmse_ratio, mean_abs,
  abs_tstat), cols = list(unit = "unit", time = "time", outcome = "outcome",
  treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{statfuncs}{Function to compute test stats}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
att estimates, test statistics, p-values
}
\description{
Get the weighted permutation distribution of SC estimate test statistics
estimating p-scores with logit-link synth
}
