% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{add_inference}
\alias{add_inference}
\title{Function to add inference to augsynth object}
\usage{
add_inference(object, inf_type = "conformal", ...)
}
\arguments{
\item{object}{augsynth object}

\item{inf_type}{Type of inference algorithm. Options are
\itemize{
 \item{"conformal"}{Conformal inference (default)}
 \item{"jackknife+"}{Jackknife+ algorithm over time periods}
 \item{"jackknife"}{Jackknife over units}
 \item{"permutation"}{Placebo permutation, raw ATT}
 \item{"permutation_rstat"}{Placebo permutation, RMSPE adjusted ATT}
}}

\item{...}{Optional arguments for inference, for more details for each `inf_type` see
\itemize{
 \item{"conformal"}{`conformal_inf`}
 \item{"jackknife+"}{`time_jackknife_plus`}
 \item{"jackknife"}{`jackknife_se_single`}
 \item{"permutation"}{`permutation_inf`}
}}
}
\description{
Function to add inference to augsynth object
}
