% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_augsynth.R
\docType{package}
\name{augsynth-package}
\alias{augsynth-package}
\title{augsynth: A package implementing the Augmented Synthetic Controls Method}
\description{
augsynth: A package implementing the Augmented Synthetic Controls Method
}
