% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth_pre.R
\name{augsynth}
\alias{augsynth}
\title{Fit Augmented SCM}
\usage{
augsynth(form, unit, time, data, t_int = NULL, ...)
}
\arguments{
\item{form}{outcome ~ treatment | auxillary covariates}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{data}{Panel data as dataframe}

\item{t_int}{Time of intervention (used for single-period treatment
only)}

\item{...}{Optional arguments
\itemize{
  \item Single period augsynth with/without multiple outcomes
    \describe{
      \item{"progfunc"}{What function to use to impute control outcomes: Ridge=Ridge regression, None=No outcome model, GSYN=gSynth}
      \item{"scm"}{Whether the SCM weighting function is used}
      \item{"fixedeff"}{Whether to include a unit fixed effect, default is FALSE }
      \item{"cov_agg"}{Covariate aggregation functions, if NULL then use mean with NAs omitted}
    }
  \item Multi period (staggered) augsynth
   \describe{
         \item{"relative"}{Whether to compute balance by relative time}
         \item{"n_leads"}{How long past treatment effects should be estimated for}
         \item{"n_lags"}{Number of pre-treatment periods to balance, default is to balance all periods}
         \item{"alpha"}{Fraction of balance for individual balance}
         \item{"lambda"}{Regularization hyperparameter, default = 0}
         \item{"force"}{Include "none", "unit", "time", "two-way" fixed effects. Default: "two-way"}
        }
}}
}
\value{
augsynth or multisynth object (depending on dispatch) that contains (among other things):
         \item{"weights"}{weights}
         \item{"data"}{Panel data as matrices}
}
\description{
The general `augsynth()` method dispatches to `single_augsynth`,
`augsynth_multiout`, or `multisynth` based on the number of
outcomes and treatment times.  See documentation for these methods
for further detail.
}
\seealso{
`single_augsynth`, `augsynth_multiout`, `multisynth`
}
