% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{augsynth}
\alias{augsynth}
\title{Fit Augmented SCM}
\usage{
augsynth(formula, unit, time, t_int, data, progfunc = "Ridge",
  weightfunc = "SCM", opts.prog = list(), opts.weights = list())
}
\arguments{
\item{formula}{outcome ~ treatment}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{t_int}{Time of intervention}

\item{data}{Panel data as dataframe}

\item{progfunc}{Outcome model to us, default is ridge regression}

\item{weightfunc}{Weighting function to use, default is SCM}

\item{opts.prog}{Optional options for fitting outcome model}

\item{opts.weights}{Optional options for fitting synth weights}
}
\value{
augsynth object that contains:
        \itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
         \item{"data"}{Panel data as matrices}
        }
}
\description{
Fit Augmented SCM
}
