% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{augsynth_multiout}
\alias{augsynth_multiout}
\title{Fit Augmented SCM with multiple outcomes}
\usage{
augsynth_multiout(
  form,
  unit,
  time,
  t_int,
  data,
  progfunc = c("Ridge", "None"),
  scm = T,
  fixedeff = FALSE,
  cov_agg = NULL,
  combine_method = "avg",
  ...
)
}
\arguments{
\item{form}{outcome ~ treatment | auxillary covariates}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{t_int}{Time of intervention}

\item{data}{Panel data as dataframe}

\item{progfunc}{What function to use to impute control outcomes
Ridge=Ridge regression,
None=No outcome model,}

\item{scm}{Whether the SCM weighting function is used}

\item{fixedeff}{Whether to include a unit fixed effect, default F}

\item{cov_agg}{Covariate aggregation functions, if NULL then use mean with NAs omitted}

\item{combine_method}{How to combine outcomes: `concat` concatenates outcomes and `avg` averages them, default: 'avg'}

\item{...}{optional arguments for outcome model}
}
\value{
augsynth object that contains:
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
         \item{"data"}{Panel data as matrices}
}
\description{
Fit Augmented SCM with multiple outcomes
}
