% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_control.R
\name{donor_table}
\alias{donor_table}
\title{Return a summary data frame donor units used in the model, along
with their RMSPEs and synthetic weights}
\usage{
donor_table(augsynth)
}
\arguments{
\item{augsynth}{Augsynth object to be plotted}
}
\description{
If the augsynth object does not have permutation-based inference
results, the function will call that form of inference, in order to
calculate the RMSPEs for each donor unit in turn.
}
