% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{double_screen}
\alias{double_screen}
\title{Screen covariates for the outcome process with LASSO and selection with
SC with infinity norm}
\usage{
double_screen(ipw_format, syn_format, avg = FALSE, by = 1)
}
\arguments{
\item{ipw_format}{Output of `format_ipw`}

\item{syn_format}{Output of `syn_format`}

\item{avg}{Fit the average post-period rather than time periods separately}

\item{by}{Step size for binary search of minimal L infinity error}
}
\value{
\itemize{
          \item{selX }{Selected covariates}
          \item{params }{Regression parameters}}
}
\description{
Screen covariates for the outcome process with LASSO and selection with
SC with infinity norm
}
