% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{double_screen}
\alias{double_screen}
\title{Screen covariates for the outcome process with LASSO and selection with
SC with infinity norm}
\usage{
double_screen(ipw_format, syn_format, type = "sep", mine = 0, by = 1)
}
\arguments{
\item{ipw_format}{Output of `format_ipw`}

\item{syn_format}{Output of `syn_format`}

\item{type}{How to fit outcome model(s)
\itemize{
 \item{sep }{Separate outcome models}
 \item{avg }{Average responses into 1 outcome}
 \item{multi }{Use multi response regression in glmnet}}}

\item{mine}{Smallest imbalance to consider, default 0}

\item{by}{Step size for binary search of minimal L infinity error}
}
\value{
\itemize{
          \item{selX }{Selected covariates}
          \item{params }{Regression parameters}}
}
\description{
Screen covariates for the outcome process with LASSO and selection with
SC with infinity norm
}
