% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{extract_covariates}
\alias{extract_covariates}
\title{Helper function to extract covariate matrix from data}
\usage{
extract_covariates(form, unit, time, t_int, data, cov_agg)
}
\description{
Helper function to extract covariate matrix from data
}
