% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{fit_augsynth_internal}
\alias{fit_augsynth_internal}
\title{Internal function to fit augmented SCM}
\usage{
fit_augsynth_internal(wide, synth_data, Z, progfunc, scm, fixedeff, ...)
}
\arguments{
\item{wide}{Data formatted from format_data}

\item{synth_data}{Data formatted from foramt_synth}

\item{Z}{Matrix of auxiliary covariates}

\item{progfunc}{outcome model to use}

\item{scm}{Whether to fit SCM}

\item{...}{Extra args for outcome model}

\item{fixed_eff}{Whether to de-mean synth}
}
\description{
Internal function to fit augmented SCM
}
