% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{fit_augsynth_multiout_internal}
\alias{fit_augsynth_multiout_internal}
\title{Internal function to fit augmented SCM with multiple outcomes}
\usage{
fit_augsynth_multiout_internal(wide_list, combine_method, Z, progfunc, scm,
  fixedeff, ...)
}
\arguments{
\item{Z}{Matrix of auxiliary covariates}

\item{progfunc}{outcome model to use}

\item{scm}{Whether to fit SCM}

\item{...}{Extra args for outcome model}

\item{wide}{Data formatted from format_data}

\item{synth_data}{Data formatted from foramt_synth}

\item{fixed_eff}{Whether to de-mean synth}
}
\description{
Internal function to fit augmented SCM with multiple outcomes
}
