% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{fit_dr_formatted}
\alias{fit_dr_formatted}
\title{Fit a regularized outcome model and synthetic controls
for a double robust estimator}
\usage{
fit_dr_formatted(data_out, alpha_w = NULL, alpha_o = NULL, net_param = 1)
}
\arguments{
\item{data_out}{formatted data from format_ipw}

\item{alpha_w}{regularization parameter for p score, if NULL use CV}

\item{alpha_o}{regularization parameter for outcome model, if NULL use CV}

\item{net_param}{elastic net parameter, 1: lasso, 0: ridge, default: 1}
}
\value{
synthetic control weights,
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Fit a regularized outcome model and synthetic controls
for a double robust estimator
}
