% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_entropy_formatted}
\alias{fit_entropy_formatted}
\title{Fit l2 entropy regularized synthetic controls
by solving the dual problem}
\usage{
fit_entropy_formatted(data_out, eps, lasso = FALSE, max_iters = 1000)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{eps}{List of bounds on synthetic control differences for each
outcome, if only one outcome type then a scalar}

\item{lasso}{Whether to do lasso (every covariate is separate)}

\item{max_iters}{Maximum number of iterations}
}
\value{
synthetic control weights
}
\description{
Fit l2 entropy regularized synthetic controls
by solving the dual problem
}
