% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{fit_ipw_formatted}
\alias{fit_ipw_formatted}
\title{Fit IPW weights with a logit propensity score model}
\usage{
fit_ipw_formatted(data_out, alpha_w)
}
\arguments{
\item{data_out}{formatted data from format_entropy}

\item{alpha_w}{regularization parameter for weights}
}
\value{
inverse of predicted propensity scores
        outcome regression parameters
        control outcomes
        treated outcomes
        boolean for treated
}
\description{
Fit IPW weights with a logit propensity score model
}
