% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{fit_prog_gsynth}
\alias{fit_prog_gsynth}
\title{Use gsynth to fit factor model for E[Y(0)|X]}
\usage{
fit_prog_gsynth(X, y, trt, opts = NULL)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}

\item{opts}{List of options for randomForest}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Use gsynth to fit factor model for E[Y(0)|X]
}
