% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{fit_prog_reg}
\alias{fit_prog_reg}
\title{Use a separate regularized regression for each post period
to fit E[Y(0)|X]}
\usage{
fit_prog_reg(
  X,
  y,
  trt,
  alpha = 1,
  lambda = NULL,
  poly_order = 1,
  type = "sep",
  ...
)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}

\item{alpha}{Mixing between L1 and L2, default: 1 (LASSO)}

\item{lambda}{Regularization hyperparameter, if null then CV}

\item{poly_order}{Order of polynomial to fit, default 1}

\item{type}{How to fit outcome model(s)
\itemize{
 \item{sep }{Separate outcome models}
 \item{avg }{Average responses into 1 outcome}
 \item{multi }{Use multi response regression in glmnet}}}

\item{...}{optional arguments for outcome model}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Use a separate regularized regression for each post period
to fit E[Y(0)|X]
}
