% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_models.R
\name{fit_prog_rf}
\alias{fit_prog_rf}
\title{Use a separate random forest regression for each post period
to fit E[Y(0)|X]}
\usage{
fit_prog_rf(X, y, trt, avg = FALSE)
}
\arguments{
\item{X}{Matrix of covariates/lagged outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Vector of treatment indicator}

\item{opts}{List of options for randomForest
\itemize{\item{avg }{Fit the average post-period rather than time periods separately}}}
}
\value{
\itemize{
          \item{y0hat }{Predicted outcome under control}
          \item{params }{Regression parameters}}
}
\description{
Use a separate random forest regression for each post period
to fit E[Y(0)|X]
}
