% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{fit_ridgeaug_formatted}
\alias{fit_ridgeaug_formatted}
\title{Ridge augmented weights (possibly with covariates)}
\usage{
fit_ridgeaug_formatted(wide_data, synth_data, Z = NULL, lambda = NULL,
  ridge = T, scm = T)
}
\arguments{
\item{wide_data}{Output of `format_data`}

\item{synth_data}{Output of `format_synth`}

\item{Z}{Matrix of covariates, default is  NULL}

\item{lambda}{Ridge hyper-parameter, if NULL use CV}

\item{ridge}{Include ridge or not}

\item{scm}{Include SCM or not}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate (zero in this case)}
         \item{"lambda"}{Value of the ridge hyperparameter}
         \item{"ridge_mhat"}{The ridge regression predictions (for estimating the bias)}
         \item{"synw"}{The synth weights(for estimating the bias)}
}
}
\description{
Ridge augmented weights (possibly with covariates)
}
