% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{fit_ridgeaug_formatted}
\alias{fit_ridgeaug_formatted}
\title{Ridge augmented weights (possibly with covariates)}
\usage{
fit_ridgeaug_formatted(
  wide_data,
  synth_data,
  Z = NULL,
  lambda = NULL,
  ridge = T,
  scm = T,
  lambda_min_ratio = 1e-08,
  n_lambda = 20,
  lambda_max = NULL,
  holdout_length = 1,
  min_1se = T,
  ...
)
}
\arguments{
\item{wide_data}{Output of `format_data`}

\item{synth_data}{Output of `format_synth`}

\item{Z}{Matrix of covariates, default is  NULL}

\item{lambda}{Ridge hyper-parameter, if NULL use CV}

\item{ridge}{Include ridge or not}

\item{scm}{Include SCM or not}

\item{lambda_min_ratio}{Ratio of the smallest to largest lambda when tuning lambda values}

\item{n_lambda}{Number of lambdas to consider between the smallest and largest lambda value}

\item{lambda_max}{Initial (largest) lambda, if NULL sets it to be (1+norm(X_1-X_c))^2}

\item{holdout_length}{Length of conseuctive holdout period for when tuning lambdas}

\item{min_1se}{If TRUE, chooses the maximum lambda within 1 standard error of the lambda that minimizes the CV error, if FALSE chooses the optimal lambda; default TRUE}

\item{...}{optional arguments for outcome model}
}
\value{
\itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate (zero in this case)}
         \item{"lambda"}{Value of the ridge hyperparameter}
         \item{"ridge_mhat"}{The ridge regression predictions (for estimating the bias)}
         \item{"synw"}{The synth weights(for estimating the bias)}
         \item{"lambdas"}{List of lambda values evaluated to tune ridge regression}
         \item{"lambda_errors"}{"The MSE associated with each lambda term in lambdas."}
         \item{"lambda_errors_se"}{"The SE of the MSE associated with each lambda term in lambdas."}
}
}
\description{
Ridge augmented weights (possibly with covariates)
}
