% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{fit_svd_formatted}
\alias{fit_svd_formatted}
\title{Fit synthetic controls on outcomes after performing SVD}
\usage{
fit_svd_formatted(data_out, r)
}
\arguments{
\item{data_out}{Panel data formatted by Synth::dataprep}

\item{r}{Rank of matrix for SVD}
}
\description{
Fit synthetic controls on outcomes after performing SVD
}
