% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{fit_synth_formatted}
\alias{fit_synth_formatted}
\title{Fit synthetic controls on outcomes after formatting data}
\usage{
fit_synth_formatted(data_out)
}
\arguments{
\item{data_out}{Panel data formatted by Synth::dataprep}
}
\description{
Fit synthetic controls on outcomes after formatting data
}
