% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_synth}
\alias{format_synth}
\title{Format program eval matrices into synth form}
\usage{
format_synth(X, trt, y)
}
\arguments{
\item{X}{Matrix of pre-treatment outcomes}

\item{trt}{Vector of treatment assignments}

\item{y}{Matrix of post-treatment outcomes}
}
\value{
List with data formatted as Synth::dataprep
}
\description{
Format program eval matrices into synth form
}
