% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Fit l2_entropy regularized synthetic controls on outcomes}
\usage{
get_entropy(outcomes, metadata, trt_unit = 1, eps = NULL,
  outcome_col = NULL, lasso = FALSE, cols = list(unit = "unit", time =
  "time", outcome = "outcome", treated = "treated"), max_iters = 1000)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps}{Bound on synthetic control differences}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}

\item{lasso}{Whether to do lasso (every covariate is separate)}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}

\item{max_iters}{Maximum number of iterations}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit l2_entropy regularized synthetic controls on outcomes
}
