% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{get_ipw}
\alias{get_ipw}
\title{Fit IPW weights with a logit propensity score model}
\usage{
get_ipw(outcomes, metadata, alpha_w = NULL, net_param = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{alpha_w}{regularization parameter for weights}

\item{net_param}{elastic net parameter, 1: lasso, 0: ridge, default: 1}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit IPW weights with a logit propensity score model
}
