% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{get_l2_entropy}
\alias{get_l2_entropy}
\title{Fit l2_entropy regularized synthetic controls on outcomes}
\usage{
get_l2_entropy(outcomes, metadata, trt_unit = 1, eps = NULL,
  outcome_col = NULL)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{eps}{Bound on synthetic control differences}

\item{outcome_col}{Column name which identifies outcomes, if NULL then
assume only one outcome}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit l2_entropy regularized synthetic controls on outcomes
}
