% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{get_synth}
\alias{get_synth}
\title{Fit synthetic controls on outcomes}
\usage{
get_synth(outcomes, metadata, trt_unit = 1, cols = list(unit = "unit", time
  = "time", outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 0}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
outcomes with additional synthetic control added and weights
}
\description{
Fit synthetic controls on outcomes
}
