% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_synth.R
\name{get_uniform}
\alias{get_uniform}
\title{Fit synthetic controls for multiple outcomes with uniform weights}
\usage{
get_uniform(outcomes, metadata, trt_unit = 1)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression), default: 1}
}
\value{
both outcomes with additional synthetic controls added
        weights for both synthetic controls
}
\description{
Fit synthetic controls for multiple outcomes with uniform weights
}
