% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{jackknife_se_single}
\alias{jackknife_se_single}
\title{Estimate standard errors for single ASCM with the jackknife
Do this for ridge-augmented synth}
\usage{
jackknife_se_single(ascm)
}
\arguments{
\item{ascm}{Fitted augsynth object}
}
\description{
Estimate standard errors for single ASCM with the jackknife
Do this for ridge-augmented synth
}
