% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highdim.R
\name{lasso_screen}
\alias{lasso_screen}
\title{Screen covariates for the outcome process}
\usage{
lasso_screen(ipw_format, syn_format, opts = list(avg = FALSE))
}
\arguments{
\item{ipw_format}{Output of `format_ipw`}

\item{syn_format}{Output of `syn_format`    r}

\item{opts}{List of options for glmnet:
\itemize{\item{avg }{Fit the average post-period rather than time periods separately}}}
}
\value{
\itemize{
          \item{selX }{Selected covariates}
          \item{params }{Regression parameters}}
}
\description{
Screen covariates for the outcome process
}
