% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{loo_se_ridgeaug}
\alias{loo_se_ridgeaug}
\title{Use leave out one estimates (placebo gaps) to estimate unit-level variance
Do this for ridge-augmented synth}
\usage{
loo_se_ridgeaug(outcomes, metadata, trt_unit = 1, lambda = NULL, scm = T,
  use_weights = T, hc = 0, cols = list(unit = "unit", time = "time",
  outcome = "outcome", treated = "treated"))
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Treated unit}

\item{lambda}{Ridge hyper-parameter, if NULL use CV}

\item{scm}{Include SCM or not}

\item{use_weights}{Whether to use weights in se estimate, default: FALSE}

\item{hc}{Type of HC variance estimate (-1 for homoskedastic)}

\item{cols}{Column names corresponding to the units,
time variable, outcome, and treated indicator}
}
\value{
att estimates, test statistics, p-values
}
\description{
Use leave out one estimates (placebo gaps) to estimate unit-level variance
Do this for ridge-augmented synth
}
