% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth_qp.R
\name{make_constraint_mats}
\alias{make_constraint_mats}
\title{Create constraint matrices for multisynth QP}
\usage{
make_constraint_mats(trt, grps, n_leads)
}
\arguments{
\item{trt}{Vector of treatment levels/times}

\item{grps}{Unique treatment time groups}

\item{n_leads}{Number of time periods after treatment to impute control values.
For units treated at time T_j, all units treated after T_j + n_leads
will be used as control values. If larger than the number of periods,
only never never treated units (pure controls) will be used as comparison units#'}
}
\value{
\itemize{
         \item{"Amat"}{Linear constraint matrix}
         \item{"lvec"}{Lower bounds for linear constraints}
         \item{"uvec"}{Upper bounds for linear constraints}
        }
}
\description{
Create constraint matrices for multisynth QP
}
