% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{multisynth}
\alias{multisynth}
\title{Fit staggered synth}
\usage{
multisynth(form, unit, time, data, relative = T, n_leads = NULL,
  alpha = NULL, lambda = 0, force = "two-way", n_factors = NULL,
  opts_weights = NULL)
}
\arguments{
\item{form}{outcome ~ treatment | auxillary covariates}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{data}{Panel data as dataframe}

\item{relative}{Whether to compute balance by relative time}

\item{n_leads}{How long past treatment effects should be estimated for}

\item{alpha}{Fraction of balance for individual balance}

\item{lambda}{Regularization hyperparameter, default = 0}

\item{force}{Include "none", "unit", "time", "two-way" fixed effects. Default: "two-way"}

\item{n_factors}{Number of factors for interactive fixed effects, default does CV}
}
\value{
augsynth object that contains:
        \itemize{
         \item{"weights"}{weights}
         \item{"data"}{Panel data as matrices}
        }
}
\description{
Fit staggered synth
}
