% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{multisynth_formatted}
\alias{multisynth_formatted}
\title{Internal funciton to fit staggered synth with formatted data}
\usage{
multisynth_formatted(wide, relative = T, n_leads = NULL,
  n_lags = NULL, alpha = NULL, lambda = 0, force = "two-way",
  n_factors = NULL, scm = T, time_w = F, lambda_t = 0,
  fit_resids = T, ...)
}
\arguments{
\item{wide}{List containing data elements}

\item{relative}{Whether to compute balance by relative time}

\item{n_leads}{How long past treatment effects should be estimated for}

\item{n_lags}{Number of pre-treatment periods to balance, default is to balance all periods}

\item{alpha}{Fraction of balance for individual balance}

\item{lambda}{Regularization hyperparameter, default = 0}

\item{force}{Include "none", "unit", "time", "two-way" fixed effects. Default: "two-way"}

\item{n_factors}{Number of factors for interactive fixed effects, default does CV}

\item{scm}{Whether to fit scm weights}

\item{time_w}{Whether to fit time weights}

\item{residuals}{Whether to fit SCM on the residuals or not}
}
\value{
augsynth object that contains:
        \itemize{
         \item{"weights"}{weights}
         \item{"data"}{Panel data as matrices}
        }
}
\description{
Internal funciton to fit staggered synth with formatted data
}
