% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_synth_qp.R
\name{multisynth_qp}
\alias{multisynth_qp}
\title{Internal function to fit synth with staggered adoption with a QP solver}
\usage{
multisynth_qp(X, trt, mask, n_leads = NULL, n_lags = NULL,
  relative = T, alpha = 0, lambda = 0)
}
\arguments{
\item{X}{Matrix of pre-final intervention outcomes, or list of such matrices after transformations}

\item{trt}{Vector of treatment levels/times}

\item{mask}{Matrix with indicators for observed pre-intervention times for each treatment group}

\item{n_leads}{Number of time periods after treatment to impute control values.
For units treated at time T_j, all units treated after T_j + n_leads
will be used as control values. If larger than the number of periods,
only never never treated units (pure controls) will be used as comparison units}

\item{n_lags}{Number of pre-treatment periods to balance, default is to balance all periods}

\item{relative}{Whether to re-index time according to treatment date, default T}

\item{alpha}{Hyper-parameter that controls trade-off between overall and individual balance.
Larger values of alpha place more emphasis on individual balance.
Balance measure is
  alpha ||global|| + (1-alpha) ||individual||
Default: 0}

\item{lambda}{Regularization hyper-parameter. Default, 0}
}
\description{
Internal function to fit synth with staggered adoption with a QP solver
}
