% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation_plot}
\alias{permutation_plot}
\title{Generate permutation plots}
\usage{
permutation_plot(augsynth, inf_type = "permutation")
}
\arguments{
\item{inf_type}{Inference type (takes a value of 'permutation' or 'permutation_rstat')
Type of inference algorithm. Inherits inf_type from `object` or otherwise defaults to "conformal". Options are
        \itemize{
         \item{"If numeric"}{A multiple of the treated unit's RMSPE above which donor units will be dropped}
         \item{"If a character"}{The name or names of donor units to be dropped based on the `unit` parameter in the augsynth model}
        }

with an RMSPE of more than `drop` times the treated units' RMSPE from the donor plot. Defaults to 20x.}

\item{results}{Results from calling augsynth()}
}
\description{
Generate permutation plots
}
