% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{plot.augsynth}
\alias{plot.augsynth}
\title{Plot function for augsynth}
\usage{
\method{plot}{augsynth}(x, ...)
}
\arguments{
\item{x}{Augsynth object to be plotted}

\item{...}{Optional arguments}
}
\description{
Plot function for augsynth
}
