% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{plot.augsynth}
\alias{plot.augsynth}
\title{Plot function for augsynth}
\usage{
\method{plot}{augsynth}(x, ci = TRUE, cv = FALSE, inf_type = NULL, ...)
}
\arguments{
\item{x}{Augsynth object to be plotted}

\item{ci}{Boolean, whether to get confidence intervals around the point estimates}

\item{cv}{If True, plot cross validation MSE against hyper-parameter, otherwise plot effects}

\item{inf_type}{Type of inference algorithm. Inherits inf_type from `object` or otherwise defaults to "conformal". Options are
\itemize{
 \item{"conformal"}{Conformal inference (default)}
 \item{"jackknife+"}{Jackknife+ algorithm over time periods}
 \item{"jackknife"}{Jackknife over units}
 \item{"permutation"}{Placebo permutation, raw ATT}
 \item{"permutation_rstat"}{Placebo permutation, RMSPE adjusted ATT}
 \item{"None"}{Return ATT Estimate only}
}}

\item{...}{Optional arguments}
}
\description{
Plot function for augsynth
}
