% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_augsynth.R
\name{plot.augsynth}
\alias{plot.augsynth}
\title{Plot function for augsynth}
\usage{
\method{plot}{augsynth}(augsynth, se = T, jackknife = T)
}
\arguments{
\item{se}{Whether to plot standard errors}

\item{jackknife}{Whether to use jackknife or weighted SEs}
}
\description{
Plot function for augsynth
}
