% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.multisynth}
\alias{plot.multisynth}
\title{Plot function for multisynth}
\usage{
\method{plot}{multisynth}(x, ...)
}
\arguments{
\item{x}{Augsynth object to be plotted}

\item{...}{Optional arguments}
}
\description{
Plot function for multisynth
}
