% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.multisynth}
\alias{plot.multisynth}
\title{Plot function for multisynth}
\usage{
\method{plot}{multisynth}(x, ...)
}
\arguments{
\item{levels}{Treatment levels to plot for, default plots for everything}

\item{se}{Whether to plot standard errors}

\item{jackknife}{Whether to compute jackknife standard errors, default T}
}
\description{
Plot function for multisynth
}
