% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.multisynth}
\alias{plot.multisynth}
\title{Plot function for multisynth}
\usage{
\method{plot}{multisynth}(multisynth, relative = NULL, levels = NULL)
}
\arguments{
\item{relative}{Whether to estimate effects for time relative to treatment}

\item{levels}{Treatment levels to plot for, default plots for everything}
}
\description{
Plot function for multisynth
}
