% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{plot.summary.augsynth}
\alias{plot.summary.augsynth}
\title{Plot function for summary function for augsynth}
\usage{
\method{plot}{summary.augsynth}(x, inf = T, ...)
}
\arguments{
\item{x}{Summary object}

\item{inf}{Boolean, whether to plot confidence intervals}

\item{...}{Optional arguments}
}
\description{
Plot function for summary function for augsynth
}
