% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{plot.summary.augsynth}
\alias{plot.summary.augsynth}
\title{Plot function for summary function for augsynth}
\usage{
\method{plot}{summary.augsynth}(x, inf_type = "conformal", ...)
}
\arguments{
\item{x}{Summary object}

\item{inf_type}{Type of inference to plot. If left NULL, inherits
from summary object.  If different from passed summary's stored
inference, will recalculate inference.}

\item{...}{Optional arguments}
}
\description{
Plot function for summary function for augsynth
}
