% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{plot.summary.augsynth}
\alias{plot.summary.augsynth}
\title{Plot function for summary function for augsynth}
\usage{
\method{plot}{summary.augsynth}(x, ...)
}
\arguments{
\item{se}{Whether to plot standard error}
}
\description{
Plot function for summary function for augsynth
}
