% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{plot.summary.augsynth_multiout}
\alias{plot.summary.augsynth_multiout}
\title{Plot function for summary function for augsynth}
\usage{
\method{plot}{summary.augsynth_multiout}(x, inf = T, ...)
}
\arguments{
\item{x}{summary.augsynth_multiout object}

\item{...}{Optional arguments, including \itemize{\item{"se"}{Whether to plot standard error}}}
}
\description{
Plot function for summary function for augsynth
}
