% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.summary.multisynth}
\alias{plot.summary.multisynth}
\title{Plot function for summary function for multisynth}
\usage{
\method{plot}{summary.multisynth}(x, ...)
}
\arguments{
\item{x}{summary object}

\item{...}{Optional arguments}
}
\description{
Plot function for summary function for multisynth
}
