% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.summary.multisynth}
\alias{plot.summary.multisynth}
\title{Plot function for summary function for multisynth}
\usage{
\method{plot}{summary.multisynth}(x, inf = T, levels = NULL, label = T, weights = FALSE, ...)
}
\arguments{
\item{x}{summary object}

\item{inf}{Whether to plot confidence intervals}

\item{levels}{Which units/groups to plot, default is every group}

\item{label}{Whether to label the individual levels}

\item{weights}{Whether to plot the weights, default = FALSE}

\item{...}{Optional arguments}
}
\description{
Plot function for summary function for multisynth
}
