% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.summary.multisynth}
\alias{plot.summary.multisynth}
\title{Plot function for summary function for multisynth}
\usage{
\method{plot}{summary.multisynth}(summ, levels = NULL, se = T)
}
\arguments{
\item{levels}{Treatment levels to plot for, default plots for everything}

\item{se}{Whether to plot standard errors}
}
\description{
Plot function for summary function for multisynth
}
