% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{plot.summary.multisynth}
\alias{plot.summary.multisynth}
\title{Plot function for summary function for multisynth}
\usage{
\method{plot}{summary.multisynth}(x, inf = T, levels = NULL, label = T, ...)
}
\arguments{
\item{x}{summary object}

\item{levels}{Which units/groups to plot, default is every group}

\item{...}{Optional arguments}

\item{se}{Whether to plot standard errors}
}
\description{
Plot function for summary function for multisynth
}
