% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_methods.R
\name{plot_augsynth_results}
\alias{plot_augsynth_results}
\title{Plot function for augsynth or summary.augsynth objects}
\usage{
plot_augsynth_results(
  augsynth,
  cv = FALSE,
  plot_type = "estimate",
  inf_type = NULL,
  ...
)
}
\arguments{
\item{augsynth}{Augsynth or summary.augsynth object to be plotted}

\item{cv}{If True, plot cross validation MSE against hyper-parameter, otherwise plot effects}

\item{plot_type}{The stylized plot type to be returned. Options include
\itemize{
  \item{"estimate"}{The ATT and 95\% confidence interval}
  \item{"estimate only"}{The ATT without a confidence interval}
  \item{"outcomes"}{The level of the outcome variable for the treated and synthetic control units.}
  \item{"outcomes raw average"}{The level of the outcome variable for the treated and synthetic control units, along with the raw average of the donor units.}
  \item{"placebo"}{The ATTs resulting from placebo tests on the donor units.}  }}

\item{inf_type}{Type of inference algorithm. Inherits inf_type from `object` or otherwise defaults to "conformal". Options are
\itemize{
 \item{"conformal"}{Conformal inference (default)}
 \item{"jackknife+"}{Jackknife+ algorithm over time periods}
 \item{"jackknife"}{Jackknife over units}
 \item{"permutation"}{Placebo permutation, raw ATT}
 \item{"permutation_rstat"}{Placebo permutation, RMSPE adjusted ATT}
 \item{"None"}{Return ATT Estimate only}
}}

\item{...}{Optional arguments}
}
\description{
Plot function for augsynth or summary.augsynth objects
}
