% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{predict.augsynth}
\alias{predict.augsynth}
\title{Get prediction of ATT or average outcome under control}
\usage{
\method{predict}{augsynth}(object, att = F, ...)
}
\arguments{
\item{object}{augsynth object}

\item{att}{If TRUE, return the ATT, if FALSE, return imputed counterfactual}

\item{...}{Optional arguments}
}
\value{
Vector of predicted post-treatment control averages
}
\description{
Get prediction of ATT or average outcome under control
}
