% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{predict.augsynth}
\alias{predict.augsynth}
\title{Get prediction of ATT or average outcome under control}
\usage{
\method{predict}{augsynth}(augsynth, type = c("att", "outcome"))
}
\arguments{
\item{augsynth}{augsynth object}
}
\value{
Vector of predicted post-treatment control averages
}
\description{
Get prediction of ATT or average outcome under control
}
