% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{predict.multisynth}
\alias{predict.multisynth}
\title{Get prediction of average outcome under control}
\usage{
\method{predict}{multisynth}(multisynth, relative = NULL, att = F)
}
\arguments{
\item{multisynth}{Fit multisynth object}

\item{relative}{Whether to aggregate estimates according to calendar or relative time}

\item{att}{Whether to estimate the ATT or the missing counterfactual}
}
\value{
Vector of predicted post-treatment control averages for each treatment group
}
\description{
Get prediction of average outcome under control
}
