% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{predict.multisynth}
\alias{predict.multisynth}
\title{Get prediction of average outcome under control or ATT}
\usage{
\method{predict}{multisynth}(object, ...)
}
\arguments{
\item{object}{Fit multisynth object}

\item{...}{Optional arguments}
}
\value{
Matrix of predicted post-treatment control outcomes for each treated unit
}
\description{
Get prediction of average outcome under control or ATT
}
