% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_augsynth.R
\name{print.summary.augsynth}
\alias{print.summary.augsynth}
\title{Print function for summary function for augsynth}
\usage{
\method{print}{summary.augsynth}(summ)
}
\description{
Print function for summary function for augsynth
}
