% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{print.summary.augsynth_multiout}
\alias{print.summary.augsynth_multiout}
\title{Print function for summary function for augsynth}
\usage{
\method{print}{summary.augsynth_multiout}(x, ...)
}
\arguments{
\item{x}{summary.augsynth_multiout object}

\item{...}{Optional arguments}
}
\description{
Print function for summary function for augsynth
}
