% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{print.summary.multisynth}
\alias{print.summary.multisynth}
\title{Print function for summary function for multisynth}
\usage{
\method{print}{summary.multisynth}(x, ...)
}
\arguments{
\item{level}{Which treatment level to show summary for, default is average}
}
\description{
Print function for summary function for multisynth
}
