% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{sep_lasso_}
\alias{sep_lasso_}
\title{Internal function that does the work of sep_lasso}
\usage{
sep_lasso_(outcomes, metadata, trt_unit, by)
}
\arguments{
\item{outcomes}{Tidy dataframe with the outcomes and meta data}

\item{metadata}{Dataframe of metadata}

\item{trt_unit}{Unit that is treated (target for regression)}

\item{by}{Step size for tolerances to try}
}
\value{
List with lowest tolerances in units of standard deviation
}
\description{
Internal function that does the work of sep_lasso
}
