% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{single_augsynth}
\alias{single_augsynth}
\title{Fit Augmented SCM}
\usage{
single_augsynth(
  form,
  unit,
  time,
  t_int,
  data,
  progfunc = "ridge",
  scm = T,
  fixedeff = FALSE,
  cov_agg = NULL,
  ...
)
}
\arguments{
\item{form}{outcome ~ treatment | auxillary covariates}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{t_int}{Time of intervention}

\item{data}{Panel data as dataframe}

\item{progfunc}{What function to use to impute control outcomes
ridge=Ridge regression (allows for standard errors),
none=No outcome model,
en=Elastic Net, RF=Random Forest, GSYN=gSynth,
mcp=MCPanel,
cits=Comparitive Interuppted Time Series
causalimpact=Bayesian structural time series with CausalImpact}

\item{scm}{Whether the SCM weighting function is used.  If FALSE, then package will fit the outcome model, but not calculate new donor weights to match pre-treatment covariates.  Instead, each donor unit will be equally weighted.  If TRUE, weights on donor pool will be calculated.}

\item{fixedeff}{Whether to include a unit fixed effect, default F}

\item{cov_agg}{Covariate aggregation functions, if NULL then use mean with NAs omitted}

\item{...}{optional arguments for outcome model}
}
\value{
augsynth object that contains:
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
         \item{"data"}{Panel data as matrices}
}
\description{
Fit Augmented SCM
}
