% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{single_augsynth}
\alias{single_augsynth}
\title{Fit Augmented SCM}
\usage{
single_augsynth(
  form,
  unit,
  time,
  t_int,
  data,
  progfunc = "ridge",
  scm = T,
  fixedeff = FALSE,
  cov_agg = NULL,
  inf_type = "None",
  ...
)
}
\arguments{
\item{form}{outcome ~ treatment | auxillary covariates}

\item{unit}{Name of unit column}

\item{time}{Name of time column}

\item{t_int}{Time of intervention}

\item{data}{Panel data as dataframe}

\item{progfunc}{What function to use to impute control outcomes
ridge=Ridge regression (allows for standard errors),
none=No outcome model,
en=Elastic Net, RF=Random Forest, GSYN=gSynth,
mcp=MCPanel,
cits=Comparitive Interuppted Time Series
causalimpact=Bayesian structural time series with CausalImpact}

\item{scm}{Whether the SCM weighting function is used}

\item{fixedeff}{Whether to include a unit fixed effect, default F}

\item{cov_agg}{Covariate aggregation functions, if NULL then use mean with NAs omitted}

\item{inf_type}{Type of inference algorithm. Options are
\itemize{
 \item{"conformal"}{Conformal inference (default)}
 \item{"jackknife+"}{Jackknife+ algorithm over time periods}
 \item{"jackknife"}{Jackknife over units}
 \item{"permutation"}{Placebo permutation, raw ATT}
 \item{"permutation_rstat"}{Placebo permutation, RMSPE adjusted ATT}
}}

\item{...}{optional arguments for outcome model}

\item{plot}{Whether or not to return a plot of the augsynth model}
}
\value{
augsynth object that contains:
        \itemize{
         \item{"weights"}{Ridge ASCM weights}
         \item{"l2_imbalance"}{Imbalance in pre-period outcomes, measured by the L2 norm}
         \item{"scaled_l2_imbalance"}{L2 imbalance scaled by L2 imbalance of uniform weights}
         \item{"mhat"}{Outcome model estimate}
         \item{"data"}{Panel data as matrices}
        }
}
\description{
Fit Augmented SCM
}
