% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{summary.augsynth}
\alias{summary.augsynth}
\title{Summary function for augsynth}
\usage{
\method{summary}{augsynth}(object, ...)
}
\arguments{
\item{object}{augsynth object}

\item{...}{Optional arguments}
}
\description{
Summary function for augsynth
}
