% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_augsynth.R
\name{summary.augsynth}
\alias{summary.augsynth}
\title{Summary function for augsynth}
\usage{
\method{summary}{augsynth}(augsynth, jackknife = T)
}
\arguments{
\item{jackknife}{Whether to use jackknife or weighted SEs}
}
\description{
Summary function for augsynth
}
