% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augsynth.R
\name{is.summary.augsynth}
\alias{is.summary.augsynth}
\title{Methods for accessing details of summary.augsynth object}
\usage{
is.summary.augsynth(x)
}
\arguments{
\item{x}{summary.augsynth result object}
}
\value{
is.summary.augsynth: TRUE if object is a augsynth object.
}
\description{
Methods for accessing details of summary.augsynth object
}
