% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{summary.augsynth_multiout}
\alias{summary.augsynth_multiout}
\title{Summary function for augsynth}
\usage{
\method{summary}{augsynth_multiout}(augsynth, se = T)
}
\arguments{
\item{se}{Whether to compute standard errors, default TRUE}
}
\description{
Summary function for augsynth
}
