% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outcomes.R
\name{summary.augsynth_multiout}
\alias{summary.augsynth_multiout}
\title{Summary function for augsynth}
\usage{
\method{summary}{augsynth_multiout}(object, inf = T, inf_type = "jackknife", ...)
}
\arguments{
\item{object}{augsynth_multiout object}

\item{...}{Optional arguments, including \itemize{\item{"se"}{Whether to plot standard error}}}
}
\description{
Summary function for augsynth
}
