% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisynth_class.R
\name{summary.multisynth}
\alias{summary.multisynth}
\title{Summary function for multisynth}
\usage{
\method{summary}{multisynth}(multisynth, relative = NULL)
}
\arguments{
\item{relative}{Whether to estimate effects for time relative to treatment}
}
\description{
Summary function for multisynth
}
