% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{update_augsynth}
\alias{update_augsynth}
\title{Return a new augsynth object with specified donor units removed}
\usage{
update_augsynth(augsynth, drop = 20)
}
\arguments{
\item{augsynth}{Augsynth object to be plotted}

\item{drop}{Drop donor units, based on pre-treatment RMSPE or unit name(s)}
}
\description{
Return a new augsynth object with specified donor units removed
}
