% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{wls_se_}
\alias{wls_se_}
\title{Compute model-based weighted least squares SE estimates of the ATT}
\usage{
wls_se_(X = NULL, y, trt, weights, pred_int, hc)
}
\arguments{
\item{X}{Matrix of pre-period outcomes}

\item{y}{Matrix of post-period outcomes}

\item{trt}{Treatment indicator}

\item{weights}{Balancing weights}

\item{pred_int}{Whether to add outcome control variance to make pred interval}

\item{hc}{Type of HC variance estimate (-1 for homoskedastic)}
}
\description{
Compute model-based weighted least squares SE estimates of the ATT
}
